### 2025-07-16 #8595 @danielskatz
👋 @minyj283 - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. It looks like you just need to change the ORCID link to just the ORCID (without the URL prefix).

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
