### 2025-07-16 62e94df
relocate stac to correct location, set server default for indoor project data processing column
