+++ backend/alembic/versions/9953a4575624_drop_directory_structure_column_add_is_.py
@@ -18,7 +18,7 @@ depends_on: str | None = None
 
 def upgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('indoor_project_data', sa.Column('is_initial_processing_completed', sa.Boolean(), nullable=False))
+    op.add_column('indoor_project_data', sa.Column('is_initial_processing_completed', sa.Boolean(), server_default=sa.text('false'), nullable=False))
     op.drop_column('indoor_project_data', 'directory_structure')
     # ### end Alembic commands ###
 
