+++ backend/app/api/api_v1/endpoints/indoor_project_data.py
@@ -3,6 +3,7 @@ import os
 from datetime import datetime, timedelta, date
 from typing import Any, Dict, List, Optional, Union, Sequence, Tuple
 
+import numpy as np
 import pandas as pd
 from fastapi import APIRouter, Depends, HTTPException, status
 from pydantic import UUID4, ValidationError
@@ -237,22 +238,11 @@ def read_indoor_project_data_plant(
             detail="Only RGB data supported at this time",
         )
 
-    # Check if we have images for this experiment
+    # Get all tar files for this project to find images
     indoor_project_files = crud.indoor_project_data.read_multi_by_id(
         db, indoor_project_id=indoor_project_id
     )
-    image_file_id: Union[UUID4, None] = None
-    image_directory_structure: Union[Dict, None] = None
-    if len(indoor_project_files) > 0:
-        for file in indoor_project_files:
-            if file.file_type == ".tar":
-                if (
-                    file.directory_structure.get("name")
-                    and file.directory_structure.get("children")
-                    and len(file.directory_structure["children"]) > 0
-                ):
-                    image_directory_structure = file.directory_structure
-                    image_file_id = file.id
+    tar_files = [file for file in indoor_project_files if file.file_type == ".tar"]
     try:
         # read "PPEW" worksheet into pandas dataframe
         ppew_df = pd.read_excel(
@@ -356,12 +346,15 @@ def read_indoor_project_data_plant(
     # pot_side_all_df.columns = pot_side_all_df.columns.str.replace(" ", "_")
     pot_side_avg_df.columns = pot_side_avg_df.columns.str.replace(" ", "_")
 
-    # convert planting date to datetime string YYYY-mm-dd HH:MM:SS
-    pot_ppew_df["planting_date"] = (
-        pot_ppew_df["planting_date"].apply(parse_date).dt.strftime("%Y-%m-%d %H:%M:%S")
-    )
+    # convert planting date to datetime object - use copy to avoid SettingWithCopyWarning
+    pot_ppew_df = pot_ppew_df.copy()
+    pot_ppew_df["planting_date"] = pot_ppew_df["planting_date"].apply(parse_date)
+    # Convert pandas Timestamp to Python datetime - use iloc to avoid FutureWarning
+    pot_ppew_df["planting_date"] = pot_ppew_df["planting_date"].iloc[0].to_pydatetime()
 
-    # convert scan date to date string YYYY-mm-dd
+    # convert scan date to date string YYYY-mm-dd - use copy to avoid SettingWithCopyWarning
+    pot_top_df = pot_top_df.copy()
+    pot_side_avg_df = pot_side_avg_df.copy()
     pot_top_df["scan_date"] = pot_top_df["scan_date"].dt.strftime("%Y-%m-%d")
     # pot_side_all_df["scan_date"] = pot_side_all_df["scan_date"].dt.strftime("%Y-%m-%d")
     pot_side_avg_df["scan_date"] = pot_side_avg_df["scan_date"].dt.strftime("%Y-%m-%d")
@@ -380,12 +373,21 @@ def read_indoor_project_data_plant(
         lambda col: col.fillna("") if col.dtype == "object" else col.fillna(-9999)
     )
 
+    # Ensure treatment is always a string
+    if "treatment" in pot_ppew_df.columns:
+        pot_ppew_df["treatment"] = pot_ppew_df["treatment"].astype(str)
+    if "treatment" in pot_top_df.columns:
+        pot_top_df["treatment"] = pot_top_df["treatment"].astype(str)
+    if "treatment" in pot_side_avg_df.columns:
+        pot_side_avg_df["treatment"] = pot_side_avg_df["treatment"].astype(str)
+
     # columns to send to client from ppew worksheet
[…]
