+++ docs/paper.bib
@@ -29,7 +29,7 @@
   number={6},
   pages={911},
   year={2022},
-  publisher={MDPI}
+  publisher={MDPI},
   doi={10.3390/f13060911}
 }
 
