### 2025-07-16 #8595 @minyj283
I'm not sure if this is the appropriate place to mention it, but if so: We believe _SkyPortal_ aligns most closely with our D2S system in terms of both architecture and user perspective. _DASF_ also shares some similarities, particularly in its frontend visual components, which use a framework similar to ours, and its API-based backend, but it is not as closely aligned as _SkyPortal_.

Additionally, we have updated the ORCIDs and added the missing DOIs, as suggested by the editorial bot.
