### 2025-07-21 19b4f39
add polymorphic project_type and project_uuid to project_members, create migration that backfills existing project_ids, add tests to verify project_type and project_uuid exist on new projects
