+++ backend/app/models/project.py
@@ -4,11 +4,12 @@ from typing import List, TYPE_CHECKING
 
 from datetime import date
 
-from sqlalchemy import Boolean, Date, DateTime, ForeignKey, String
+from sqlalchemy import and_, Boolean, Date, DateTime, ForeignKey, String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base_class import Base
+from app.models.project_type import ProjectType
 
 
 if TYPE_CHECKING:
@@ -47,8 +48,16 @@ class Project(Base):
         DateTime(timezone=True), nullable=True
     )
 
+    uas_members: Mapped[List["ProjectMember"]] = relationship(
+        "ProjectMember",
+        back_populates="uas_project",
+        cascade="all, delete-orphan",
+        primaryjoin="and_(ProjectMember.project_type == 'PROJECT', ProjectMember.project_uuid == Project.id)",
+        foreign_keys="[ProjectMember.project_uuid]",
+    )
+    # TODO: Remove this relationship after migration
     members: Mapped[List["ProjectMember"]] = relationship(
-        back_populates="project", cascade="all, delete"
+        back_populates="project", cascade="all, delete-orphan"
     )
     modules: Mapped[List["ProjectModule"]] = relationship(back_populates="project")
     location: Mapped["Location"] = relationship(back_populates="project")
