### 2025-07-21 8b7bd93
update api and crud to use project_type and project_uuid instead of project_id, update tests
