### 2025-07-23 a1224fc
add basic stateless refresh tokens, consolidate token validation and payload decoding
