+++ backend/app/core/security.py
@@ -2,21 +2,27 @@ import base64
 import hmac
 import hashlib
 import time
+import os
+import logging
 from datetime import datetime, timedelta, timezone
-from typing import Any, cast, Dict, Optional, Tuple
+from typing import Any, cast, Dict, Optional, Tuple, Literal
+from uuid import UUID
 
-from fastapi import Request, status
+from fastapi import Request, Response, status, HTTPException
 from fastapi.exceptions import HTTPException
 from fastapi.openapi.models import OAuthFlows as OAuthFlowsModel
 from fastapi.security import OAuth2
 from fastapi.security.utils import get_authorization_scheme_param
 from jose import jwt
+from jose.exceptions import ExpiredSignatureError
 from passlib.context import CryptContext
 from passlib.hash import sha256_crypt
+from sqlalchemy.orm import Session
 
 from app.core.config import settings
 from app.models.single_use_token import SingleUseToken
 
+logger = logging.getLogger("__name__")
 
 pwd_context = CryptContext(
     schemes=["bcrypt"],
@@ -32,9 +38,25 @@ def create_access_token(subject: str | Any, expire: datetime | None = None) -> s
         expire = datetime.now(timezone.utc) + timedelta(
             minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES
         )
-    to_encode = {"exp": expire, "sub": str(subject)}
+    to_encode = {"exp": expire, "sub": str(subject), "type": "access"}
     encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=ALGORITHM)
     return encoded_jwt
+
+
+def create_refresh_token(subject: str | Any, expire: datetime | None = None) -> str:
+    """Create JWT refresh token with expiration defined in settings."""
+    if not expire:
+        expire = datetime.now(timezone.utc) + timedelta(
+            days=settings.REFRESH_TOKEN_EXPIRE_DAYS
+        )
+    to_encode = {"exp": expire, "sub": str(subject), "type": "refresh"}
+    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=ALGORITHM)
+    return encoded_jwt
+
+
+def decode_token(token: str) -> dict[str, Any]:
+    """Decode JWT token and return payload."""
+    return jwt.decode(token, settings.SECRET_KEY, algorithms=[ALGORITHM])
 
 
 def sign_map_tile_payload(payload: str, expiration: int = 600) -> Tuple[str, int]:
@@ -80,6 +102,127 @@ def get_token_hash(token: str, salt: str) -> str:
 def check_token_expired(token: SingleUseToken, minutes: int = 60) -> bool:
     """Check if token is older than value passed in minutes (defaults to 1 hour)."""
     return datetime.now(tz=timezone.utc) > (token.created_at + timedelta(minutes))
+
+
+def validate_token_and_get_payload(token: str, expected_type: str) -> dict[str, Any]:
+    """Validate JWT token, check type, and return payload with proper error handling.
+
+    Args:
+        token: JWT token string
+        expected_type: Expected token type ("access" or "refresh")
+
+    Returns:
+        dict: Token payload
+
+    Raises:
+        HTTPException: For various token validation failures
+    """
+    # Decode token and handle expiration specifically
+    try:
+        payload = decode_token(token)
+    except ExpiredSignatureError:
+        token_type_title = expected_type.title()
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail=f"{token_type_title} token expired",
+        )
+    except Exception:
+        token_type_title = expected_type.title()
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail=f"Invalid {expected_type} token",
+        )
+
[…]
