+++ backend/app/api/api_v1/endpoints/auth.py
@@ -1,7 +1,7 @@
 import logging
 import os
 import secrets
-from typing import Any, Annotated
+from typing import Any, Annotated, Literal
 
 from fastapi import (
     APIRouter,
@@ -13,6 +13,7 @@ from fastapi import (
     Response,
     status,
 )
+from fastapi import Cookie
 from fastapi.responses import RedirectResponse
 from fastapi.security import OAuth2PasswordRequestForm
 from pydantic import EmailStr
@@ -20,7 +21,6 @@ from sqlalchemy.orm import Session
 
 from app import crud, models, schemas
 from app.api import deps, mail
-from app.api.utils import str_to_bool
 from app.core import security
 from app.core.config import settings
 
@@ -54,27 +54,48 @@ def login_access_token(
             status_code=status.HTTP_403_FORBIDDEN, detail="Account requires approval"
         )
     access_token = security.create_access_token(user.id)
+    refresh_token = security.create_refresh_token(user.id)
 
-    # Toggle secure off if running tests
-    secure_cookie = True
-    samesite = "none"
-    try:
-        if str_to_bool(os.environ.get("RUNNING_TESTS", False)) or not str_to_bool(
-            os.environ.get("HTTP_COOKIE_SECURE", True)
-        ):
-            secure_cookie = False
-            samesite = "lax"
-    except ValueError:
-        logger.exception("Defaulting to secure cookie")
+    # Set authentication cookies
+    security.set_auth_cookies(response, access_token, refresh_token)
 
-    response.set_cookie(
-        key="access_token",
-        value=f"Bearer {access_token}",
-        httponly=True,
-        secure=secure_cookie,
-        samesite=samesite,
-    )
     return status.HTTP_200_OK
+
+
+@router.post("/refresh-token")
+def refresh_access_token(
+    response: Response,
+    refresh_token: str = Cookie(None),
+    db: Session = Depends(deps.get_db),
+) -> Any:
+    """Refresh access token using refresh token."""
+    # Check if refresh token is provided
+    if not refresh_token:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED, detail="Refresh token not found"
+        )
+    # Check if refresh token starts with "Bearer "
+    if not refresh_token.startswith("Bearer "):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid refresh token"
+        )
+    # Remove "Bearer " prefix and strip whitespace
+    token = refresh_token.removeprefix("Bearer ").strip()
+
+    # Validate token and get payload
+    payload = security.validate_token_and_get_payload(token, "refresh")
+
+    # Get user from payload
+    user = security.get_user_from_token_payload(db, payload)
+
+    # Issue new access token
+    new_access_token = security.create_access_token(user.id)
+    new_refresh_token = security.create_refresh_token(user.id)
+
+    # Set authentication cookies
+    security.set_auth_cookies(response, new_access_token, new_refresh_token)
+
+    return {"msg": "Access token refreshed"}
 
 
 @router.get("/remove-access-token")
