+++ backend/app/schemas/__init__.py
@@ -31,6 +31,7 @@ from .project_like import ProjectLike, ProjectLikeCreate, ProjectLikeUpdate
 from .project_member import ProjectMember, ProjectMemberCreate, ProjectMemberUpdate
 from .project_module import ProjectModule, ProjectModuleCreate, ProjectModuleUpdate
 from .raw_data import RawData, RawDataCreate, RawDataUpdate, RawDataMetadata
+from .refresh_token import RefreshToken, RefreshTokenCreate, RefreshTokenUpdate
 from .shortened_url import (
     ShortenedUrl,
     ShortenedUrlCreate,
