+++ backend/app/crud/__init__.py
@@ -15,6 +15,7 @@ from .crud_project_like import project_like
 from .crud_project_member import project_member
 from .crud_project_module import project_module
 from .crud_raw_data import raw_data
+from .crud_refresh_token import refresh_token
 from .crud_shortened_url import shortened_url
 from .crud_team import team
 from .crud_team_member import team_member
