### 2025-07-24 62650ba
update label weight hierarchy
