### 2025-07-31 69fe87c
fallback for compound crs in chm generation
