### 2025-08-05 241765c
add support for panoramic images
