### 2025-08-12 2dd5a81
make uploaded shapefile and geojson editable with geoman, improve geoman handles
