### 2025-08-12 1247f50
fix outdated tests
