+++ backend/app/tests/api/api_v1/test_vector_layers.py
@@ -49,7 +49,7 @@ def test_read_vector_layer_with_manager_owner_role(
     project = create_project(db)
     point_feature_collection = create_feature_collection(db, "point", project.id)
     create_project_member(
-        db, role=Role.MANAGER, member_id=current_user.id, project_id=project.id
+        db, role=Role.MANAGER, member_id=current_user.id, project_uuid=project.id
     )
     vector_layer_feature_id = point_feature_collection.features[0].properties[
         "feature_id"
@@ -80,7 +80,7 @@ def test_read_vector_layer_with_project_viewer_role(
     project = create_project(db)
     point_feature_collection = create_feature_collection(db, "point", project.id)
     create_project_member(
-        db, role=Role.VIEWER, member_id=current_user.id, project_id=project.id
+        db, role=Role.VIEWER, member_id=current_user.id, project_uuid=project.id
     )
     vector_layer_feature_id = point_feature_collection.features[0].properties[
         "feature_id"
@@ -177,7 +177,7 @@ def test_read_vector_layers_with_project_manager_role(
     fc2 = create_feature_collection(db, "linestring", project_id=project.id)
     fc3 = create_feature_collection(db, "polygon", project_id=project.id)
     create_project_member(
-        db, role=Role.MANAGER, member_id=current_user.id, project_id=project.id
+        db, role=Role.MANAGER, member_id=current_user.id, project_uuid=project.id
     )
 
     response = client.get(f"{settings.API_V1_STR}/projects/{project.id}/vector_layers")
@@ -203,7 +203,7 @@ def test_read_vector_layers_with_project_viewer_role(
     fc2 = create_feature_collection(db, "linestring", project_id=project.id)
     fc3 = create_feature_collection(db, "polygon", project_id=project.id)
     create_project_member(
-        db, role=Role.VIEWER, member_id=current_user.id, project_id=project.id
+        db, role=Role.VIEWER, member_id=current_user.id, project_uuid=project.id
     )
 
     response = client.get(f"{settings.API_V1_STR}/projects/{project.id}/vector_layers")
@@ -290,7 +290,7 @@ def test_remove_vector_layer_with_project_manager_role(
     current_user = get_current_user(db, normal_user_access_token)
     project = create_project(db)
     create_project_member(
-        db, role=Role.MANAGER, member_id=current_user.id, project_id=project.id
+        db, role=Role.MANAGER, member_id=current_user.id, project_uuid=project.id
     )
     feature_collection = create_feature_collection(
         db, geom_type="multipoint", project_id=project.id
@@ -314,7 +314,7 @@ def test_remove_vector_layer_with_project_viewer_role(
     current_user = get_current_user(db, normal_user_access_token)
     project = create_project(db)
     create_project_member(
-        db, role=Role.VIEWER, member_id=current_user.id, project_id=project.id
+        db, role=Role.VIEWER, member_id=current_user.id, project_uuid=project.id
     )
     feature_collection = create_feature_collection(
         db, geom_type="multipoint", project_id=project.id
