### 2025-08-12 b03d257
optional parameters for limiting the color map options, better validation for x & y axis
