### 2025-08-12 1a5a9cb
use radio buttons for changing camera orientation, finish updating indoor api endpoints
