### 2025-08-13 0b40b15
split group_by into two sets of radio buttons
