+++ frontend/src/components/pages/workspace/indoorProjects/PotGroupModule/validationSchema.ts
@@ -4,18 +4,10 @@ const validationSchema = Yup.object({
   cameraOrientation: Yup.string()
     .oneOf(['top', 'side'], 'Invalid value')
     .required('Required field'),
-  groupBy: Yup.string()
-    .oneOf(
-      [
-        'treatment',
-        'description',
-        'treatment_description',
-        'all_pots',
-        'single_pot',
-      ],
-      'Invalid value'
-    )
+  plottedBy: Yup.string()
+    .oneOf(['groups', 'pots'], 'Invalid value')
     .required('Required field'),
+  accordingTo: Yup.string().required('Required field'),
 });
 
 export default validationSchema;

[…]
