### 2025-08-13 1ee3989
update according to drop-down to include pot barcodes
