### 2025-08-13 3881277
remove duplicate command to drop the unique_to_project constraint, update pot module viz to retain context on page load such as selectedDAP
