### 2025-08-13 e3b5556
add download csv button for pot module group viz
