### 2025-08-14 2ec93ae
refactor plant detail component
