### 2025-08-14 a90dad9
speed up plant detail endpoint by indexing image directory once and bulk attaching images
