### 2025-08-14 ed08715
update pot module symbols and legend to be more consistent
