### 2025-08-14 fd718ce
add custom symbols to scatter plot
