+++ frontend/src/components/pages/workspace/indoorProjects/TraitModule/TraitScatterModule/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TraitScatterModuleDataVisualization';
