### 2025-08-15 ba75c80
add endpoints/crud/tests for indoor project members
