+++ backend/app/tests/crud/test_project_member.py
@@ -7,6 +7,7 @@ from app.models.project_type import ProjectType
 from app.schemas.project_member import ProjectMemberCreate, ProjectMemberUpdate
 from app.schemas.role import Role
 from app.tests.utils.project import create_project
+from app.tests.utils.indoor_project import create_indoor_project
 from app.tests.utils.project_member import create_project_member
 from app.tests.utils.team import create_team
 from app.tests.utils.team_member import create_team_member
@@ -252,3 +253,235 @@ def test_project_member_target_project_property(db: Session) -> None:
         assert target_project is not None
         assert target_project.id == project.id
         assert target_project.title == project.title
+
+
+# Indoor Project Member Tests
+def test_create_indoor_project_member(db: Session) -> None:
+    """Test creating a project member for an indoor project."""
+    project_owner = create_user(db)
+    indoor_project = create_indoor_project(db, owner_id=project_owner.id)
+    user = create_user(db)
+    project_member = create_project_member(
+        db,
+        member_id=user.id,
+        project_uuid=indoor_project.id,
+        project_type=ProjectType.INDOOR_PROJECT,
+    )
+    assert project_member
+    assert user.id == project_member.member_id
+    assert indoor_project.id == project_member.project_uuid
+    assert project_member.project_type == ProjectType.INDOOR_PROJECT
+    assert project_member.project_uuid == indoor_project.id
+    assert project_member.role == Role.VIEWER  # default role
+
+
+def test_create_indoor_project_member_with_email(db: Session) -> None:
+    """Test creating a project member for an indoor project using email."""
+    project_owner = create_user(db)
+    indoor_project = create_indoor_project(db, owner_id=project_owner.id)
+    user = create_user(db)
+
+    project_member_in = ProjectMemberCreate(email=user.email, role=Role.MANAGER)
+    project_member = crud.project_member.create_with_project(
+        db,
+        obj_in=project_member_in,
+        project_uuid=indoor_project.id,
+        project_type=ProjectType.INDOOR_PROJECT,
+    )
+
+    assert project_member
+    assert user.id == project_member.member_id
+    assert indoor_project.id == project_member.project_uuid
+    assert project_member.project_type == ProjectType.INDOOR_PROJECT
+    assert project_member.role == Role.MANAGER
+
+
+def test_create_multi_indoor_project_members(db: Session) -> None:
+    """Test creating multiple project members for an indoor project."""
+    project_owner = create_user(db)
+    indoor_project = create_indoor_project(db, owner_id=project_owner.id)
+    user1 = create_user(db)
+    user2 = create_user(db)
+
+    new_members = [(user1.id, Role.VIEWER), (user2.id, Role.MANAGER)]
+    project_members = crud.project_member.create_multi_with_project(
+        db,
+        new_members=new_members,
+        project_uuid=indoor_project.id,
+        project_type=ProjectType.INDOOR_PROJECT,
+    )
+
+    assert len(project_members) == 3  # owner + two newly created members
+    member_ids = [pm.member_id for pm in project_members]
+    assert user1.id in member_ids
+    assert user2.id in member_ids
+
+    for pm in project_members:
+        assert pm.project_uuid == indoor_project.id
+        assert pm.project_type == ProjectType.INDOOR_PROJECT
+
+
+def test_get_indoor_project_member(db: Session) -> None:
+    """Test getting a project member by indoor project and member ID."""
+    project_owner = create_user(db)
+    indoor_project = create_indoor_project(db, owner_id=project_owner.id)
+    user = create_user(db)
+
+    project_member = create_project_member(
+        db,
+        member_id=user.id,
+        project_uuid=indoor_project.id,
+        project_type=ProjectType.INDOOR_PROJECT,
+    )
+
+    retrieved_member = crud.project_member.get_by_project_and_member_id(
[…]
