### 2025-08-18 c2d26f6
allow api key for breedbase related endpoints
