### 2025-08-19 c70e449
return team objects with projects to home page map
