### 2025-08-19 b531042
display project creator name next to project title
