### 2025-08-19 #8595 @danielskatz
As this was moved, I am no longer the track chair. @openjournals/ese-eics is now
