### 2025-08-25 db63001
setup backand and frontend for handling 3DGS (.ply) data type
