+++ backend/app/utils/tusd/post_processing.py
@@ -13,6 +13,7 @@ from app.api.utils import get_data_product_dir
 from app.schemas.job import State, Status
 from app.tasks.post_upload_tasks import generate_point_cloud_preview
 from app.tasks.upload_tasks import (
+    upload_3dgs,
     upload_geotiff,
     upload_panoramic,
     upload_point_cloud,
@@ -31,6 +32,7 @@ SUPPORTED_EXTENSIONS = {
     ".png",
     ".webp",
     ".avif",
+    ".ply",
 }
 
 
@@ -131,6 +133,11 @@ def process_data_product_uploaded_to_tusd(
     elif dtype == "panoramic":
         # start panoramic process in background
         upload_panoramic.apply_async(
+            args=(str(storage_path), destination_filepath, job.id, data_product.id),
+        )
+    elif dtype == "3dgs":
+        # start 3D Gaussian Splatting process in background
+        upload_3dgs.apply_async(
             args=(str(storage_path), destination_filepath, job.id, data_product.id),
         )
     else:
