### 2025-08-26 d3f9962
add playcanvas viewer for rendering ply, update ui to handle new 3dgs data type
