### 2025-08-26 c610af2
drop unnecessary cors changes
