### 2025-08-27 224bfab
fix STAC metadata persistence issues on page refresh and form submission
