+++ backend/app/api/api_v1/endpoints/tusd.py
@@ -2,28 +2,92 @@ import os
 from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any
-from uuid import UUID
 
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
+from urllib.parse import unquote
 
 from app import crud, models
 from app.api import deps
+from app.core import security
+from urllib.parse import unquote
 from app.utils.tusd.post_processing import (
     process_data_product_uploaded_to_tusd,
     process_raw_data_uploaded_to_tusd,
 )
-from app.schemas import TUSDHook, UploadCreate, UploadUpdate
+from app.schemas import TUSDHook, UploadUpdate
 
 
 router = APIRouter()
+
+
+def _extract_access_token_from_cookie_headers(
+    cookie_headers: list[str] | None,
+) -> str | None:
+    """Extract the access token from forwarded Cookie headers.
+
+    Parses the list of Cookie headers forwarded by tusd, finds the
+    "access_token" entry, URL-decodes the value, removes optional
+    surrounding quotes, and strips the leading "Bearer " prefix if present.
+
+    Args:
+        cookie_headers (list[str] | None): List of Cookie header strings
+            forwarded by tusd (may be None).
+
+    Returns:
+        str | None: The raw JWT access token without the "Bearer " prefix
+            if extraction succeeds; otherwise None.
+    """
+    if not cookie_headers:
+        return None
+    token_value: str | None = None
+    for cookie_str in cookie_headers:
+        parts = [c.strip() for c in cookie_str.split(";")]
+        for part in parts:
+            if part.startswith("access_token="):
+                token_value = part.split("=", 1)[1]
+                break
+        if token_value:
+            break
+    if not token_value:
+        return None
+    # URL-decode and sanitize
+    token_value = unquote(token_value).strip()
+    if token_value.startswith('"') and token_value.endswith('"'):
+        token_value = token_value[1:-1]
+    if token_value.startswith("Bearer "):
+        token_value = token_value[len("Bearer ") :]
+    return token_value
+
+
+def _get_approved_user_from_token(db: Session, token: str) -> models.User:
+    """Validate access token, resolve user, and verify account is approved.
+
+    Validates the provided JWT access token, loads the associated user
+    from the database, and verifies that the account meets approval
+    requirements (email confirmed, approved, etc.).
+
+    Args:
+        db (Session): Database session used for lookups.
+        token (str): Raw JWT access token (without the "Bearer " prefix).
+
+    Returns:
+        models.User: The approved user associated with the token.
+
+    Raises:
+        HTTPException: If the token is invalid or expired, if the user
+            cannot be found, or if the account fails approval checks.
+    """
+    token_payload = security.validate_token_and_get_payload(token, "access")
+    user = security.get_user_from_token_payload(db, token_payload)
+    approved_user = deps.verify_user_account(user)
+    return approved_user
 
 
 @router.post("", status_code=status.HTTP_202_ACCEPTED)
 def handle_tusd_http_hooks(
     payload: TUSDHook,
     db: Session = Depends(deps.get_db),
-    current_user: models.User = Depends(deps.get_current_approved_user),
 ) -> Any:
     """Receives http hook requests from tusd server. Starts process for moving data
     from tusd server storage to backend static file directory.
@@ -31,7 +95,6 @@ def handle_tusd_http_hooks(
     Args:
         payload (TUSDHook): Request from tusd.
         db (Session, optional): Database session. Defaults to Depends(deps.get_db).
-        current_user (models.User, optional): User that innitiated upload to tusd. Defaults to Depends(deps.get_current_approved_user).
 
     Raises:
[…]
