### 2025-09-03 dbe1fe4
move potree viewer to backend, add cesium basemap when crs available, adjust z offset
