+++ backend/app/api/extras.py
@@ -1,11 +1,27 @@
-from typing import Any
+# extras.py
+from typing import Any, Optional, Tuple
+import json
+import subprocess
 
-from fastapi import APIRouter, Depends, HTTPException
-from fastapi.responses import RedirectResponse
+from fastapi import APIRouter, Depends, HTTPException, Query
+from fastapi.responses import HTMLResponse, RedirectResponse
 from sqlalchemy.orm import Session
 
+# Your app imports
 from app import crud
 from app.api import deps
+
+# Geospatial + LAS
+try:
+    import laspy
+except Exception:  # pragma: no cover
+    laspy = None
+
+try:
+    from pyproj import CRS, Transformer
+except Exception:  # pragma: no cover
+    CRS = None
+    Transformer = None
 
 extra_router = APIRouter()
 
@@ -25,3 +41,666 @@ async def redirect_short_url(short_id: str, db: Session = Depends(deps.get_db)) 
         url=original_url,
         status_code=303,
     )
+
+
+def pdal_percentile_z(
+    copc_path: str, pct: int = 2, decimation: int = 50
+) -> Optional[float]:
+    """
+    Fast near-ground proxy: low percentile of Z from a decimated sample.
+    Works for URLs (HTTP range) or local files.
+    """
+    pipeline = {
+        "pipeline": [
+            {"type": "readers.copc", "filename": copc_path},
+            {"type": "filters.decimation", "step": decimation},
+            {"type": "filters.stats", "dimensions": "Z", "percentile": [pct]},
+        ]
+    }
+    out = subprocess.check_output(
+        ["pdal", "pipeline", "--stdin", "--metadata"],
+        input=json.dumps(pipeline),
+        text=True,
+    )
+    meta = json.loads(out)["metadata"]
+    stats_nodes = [v for k, v in meta.items() if "filters.stats" in k]
+    if not stats_nodes:
+        return None
+    s = stats_nodes[0]["statistic"]
+    # use the percentile if present, else minimum
+    z_low = s.get(f"percentile{pct}", s.get("minimum"))
+    return float(z_low) if z_low is not None else None
+
+
+def probe_crs_and_center(
+    copc_path: str,
+) -> Tuple[Optional[str], Optional[Tuple[float, float, float]], Optional[str]]:
+    """
+    Attempt to read the COPC's CRS and compute a geographic center for Cesium.
+
+    Returns:
+      proj4_str: str | None          -> proj4 for the source CRS (for proj4js)
+      geo_center: (lon, lat, z) | None  -> center in EPSG:4326 degrees + meters height
+      diag: str | None               -> diagnostic message (for logging / UI)
+    """
+    if laspy is None or CRS is None or Transformer is None:
+        return None, None, "Missing laspy/pyproj on server"
+
+    try:
+        with laspy.open(copc_path) as f:
+            hdr = f.header
+            crs = hdr.parse_crs()  # pyproj.CRS or None
+
+            if crs is None:
+                return None, None, "No CRS in COPC header"
+
+            # Center from header mins/maxs (no point read required)
+            x0, y0, z0 = hdr.mins
+            x1, y1, z1 = hdr.maxs
+            cx = (x0 + x1) / 2.0
+            cy = (y0 + y1) / 2.0
+            cz = (z0 + z1) / 2.0
+
+            # Transform to WGS84 lon/lat (+ keep Z as-is; note height datum caveat)
+            wgs84 = CRS.from_epsg(4326)
+            to_geo = Transformer.from_crs(crs, wgs84, always_xy=True)
+            lon, lat, h = to_geo.transform(cx, cy, cz)
+
+            proj4_str = crs.to_proj4()
[…]
