### 2025-09-04 af9f205
enforce geographic coordinates for geosjon coordinates on both backend and frontend
