### 2025-09-05 c954554
update sharepotree viewer to work with new potree backend
