+++ backend/app/api/extras.py
@@ -1,15 +1,18 @@
 # extras.py
 from typing import Any, Optional, Tuple
 import json
+import os
 import subprocess
 
 from fastapi import APIRouter, Depends, HTTPException, Query
 from fastapi.responses import HTMLResponse, RedirectResponse
+from pydantic import UUID4
 from sqlalchemy.orm import Session
 
 # Your app imports
 from app import crud
 from app.api import deps
+from app.core.config import settings
 
 # Geospatial + LAS
 try:
@@ -20,8 +23,8 @@ except Exception:  # pragma: no cover
 try:
     from pyproj import CRS, Transformer
 except Exception:  # pragma: no cover
-    CRS = None
-    Transformer = None
+    CRS = None  # type: ignore
+    Transformer = None  # type: ignore
 
 extra_router = APIRouter()
 
@@ -112,24 +115,23 @@ def probe_crs_and_center(
         return None, None, f"CRS probe failed: {exc}"
 
 
-@extra_router.get("/potree-viewer", response_class=HTMLResponse)
-async def get_potree_viewer(
-    copc_path: str = Query(..., description="Path to the COPC file"),
-    is_mobile: bool = Query(
-        ...,
-        description="Whether the viewer is on a mobile device",
-    ),
-) -> HTMLResponse:
+def generate_potree_viewer_html(
+    copc_path: str,
+    is_mobile: bool,
+) -> str:
     """
-    Serves the Potree + optional Cesium viewer HTML content dynamically.
+    Generate the HTML content for the Potree + optional Cesium viewer.
 
-    Behavior:
-      - If COPC has a CRS, enable Cesium and camera sync, and inject the proj4 string + WGS84 center.
-      - If no CRS, disable Cesium and show Potree-only with a gentle banner.
+    Args:
+        copc_path: Path to the COPC file
+        is_mobile: Whether the viewer is on a mobile device
+
+    Returns:
+        HTML content as string
     """
     # Probe the CRS and dataset geographic center
     proj4_str, geo_center, diag = probe_crs_and_center(
-        copc_path.replace("http://localhost:8000", "")
+        copc_path.replace(settings.API_DOMAIN, "")
     )
     has_crs = proj4_str is not None and geo_center is not None
 
@@ -175,7 +177,7 @@ async def get_potree_viewer(
 
     JS_Z_LOW = "null" if z_low is None else f"{z_low:.3f}"
 
-    html_content = f"""<!DOCTYPE html>
+    return f"""<!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
@@ -703,4 +705,188 @@ async def get_potree_viewer(
   </body>
 </html>"""
 
+
+@extra_router.get("/potree-viewer", response_class=HTMLResponse)
+async def get_potree_viewer(
+    copc_path: str = Query(..., description="Path to the COPC file"),
+    is_mobile: bool = Query(
+        ...,
+        description="Whether the viewer is on a mobile device",
+    ),
+) -> HTMLResponse:
+    """
+    Serves the Potree + optional Cesium viewer HTML content dynamically.
+
+    Behavior:
+      - If COPC has a CRS, enable Cesium and camera sync, and inject the proj4 string + WGS84 center.
+      - If no CRS, disable Cesium and show Potree-only with a gentle banner.
+    """
+    html_content = generate_potree_viewer_html(copc_path, is_mobile)
+    return HTMLResponse(content=html_content)
+
+
+@extra_router.get("/share-potree-viewer", response_class=HTMLResponse)
+async def get_share_potree_viewer(
+    file_id: UUID4 = Query(..., description="ID of the data product file"),
+    is_mobile: bool = Query(
+        ...,
[…]
