### 2025-09-05 2d16ee3
revise ui potree error page
