+++ backend/app/models/project.py
@@ -4,12 +4,12 @@ from typing import List, TYPE_CHECKING
 
 from datetime import date
 
-from sqlalchemy import and_, Boolean, Date, DateTime, ForeignKey, String
+from sqlalchemy import Boolean, Date, DateTime, ForeignKey, String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base_class import Base
-from app.models.project_type import ProjectType
+from app.models.utils.utcnow import utcnow
 
 
 if TYPE_CHECKING:
@@ -44,6 +44,15 @@ class Project(Base):
     )
     is_active: Mapped[bool] = mapped_column(Boolean, default=True, nullable=False)
     is_published: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
+    created_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True), server_default=utcnow(), nullable=False
+    )
+    updated_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True),
+        server_default=utcnow(),
+        onupdate=utcnow(),
+        nullable=False,
+    )
     deactivated_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), nullable=True
     )
