### 2025-09-11 27137ba
conditionally register endpoints, skip tests for disabled endpoints
