### 2025-09-12 ce9b0f3
improve RPC client reliability and connection handling
