### 2025-09-22 7064c8f
add option for enabling opentelemetry for fastapi backend and titiler
