+++ docker-compose.yml
@@ -1,264 +1,0 @@
-services:
-  frontend:
-    build:
-      context: ./frontend
-      dockerfile: frontend.dockerfile
-    image: d2s-app:dev
-    platform: linux/x86_64
-    command: yarn run dev --host 0.0.0.0 --port 8080 --force
-    volumes:
-      - ./frontend/:/app
-      - /app/node_modules
-    restart: always
-    depends_on:
-      - backend
-    environment:
-      - VITE_STAC_ENABLED=true
-    env_file:
-      - frontend.env
-
-  backend:
-    build:
-      context: ./backend
-      dockerfile: backend.dockerfile
-      args:
-        INSTALL_DEV: true
-    image: d2s-api:dev
-    platform: linux/x86_64
-    volumes:
-      - ./backend:/app
-      - external-data:/srv/external-data
-      - log-data:/app/logs
-      - tusd-data:/srv/tusd-data/data
-      - user-data:/static
-    restart: always
-    depends_on:
-      db:
-        condition: service_healthy
-    environment:
-      - TILE_SIGNING_SECRET_KEY=${TILE_SIGNING_SECRET_KEY}
-      - OTEL_EXPORTER_OTLP_TRACES_ENDPOINT=http://otel-collector:4318
-    env_file:
-      - backend.env
-      - db.env
-
-  db:
-    image: postgis/postgis:15-3.3
-    platform: linux/x86_64
-    volumes:
-      - db-data:/var/lib/postgresql/data
-    restart: always
-    healthcheck:
-      test:
-        [
-          "CMD-SHELL",
-          "pg_isready --username=$$POSTGRES_USER --dbname=$$POSTGRES_DB",
-        ]
-      interval: 10s
-      timeout: 5s
-      retries: 5
-    environment:
-      - POSTGRES_USER=d2s
-      - POSTGRES_DB=d2s
-      - PGTZ=Etc/UTC
-    env_file:
-      - db.env
-
-  pgadmin:
-    image: dpage/pgadmin4
-    platform: linux/x86_64
-    restart: always
-    depends_on:
-      db:
-        condition: service_healthy
-    environment:
-      - PGADMIN_DEFAULT_EMAIL=youremail@example.com
-      - PGADMIN_DEFAULT_PASSWORD=password
-
-  redis:
-    image: redis:alpine
-    platform: linux/x86_64
-    restart: always
-
-  celery_worker:
-    image: d2s-api:dev
-    platform: linux/x86_64
-    command: /app/celery/worker-start.sh
-    volumes:
-      - external-data:/srv/external-data
-      - tusd-data:/srv/tusd-data/data
-      - user-data:/static
-    restart: always
-    depends_on:
-      - redis
-      - db
-    env_file:
-      - backend.env
-      - db.env
-    environment:
-      - CELERY_WORKERS=4
-      - TILE_SIGNING_SECRET_KEY=${TILE_SIGNING_SECRET_KEY}
-
-  celery_beat:
-    image: d2s-api:dev
-    platform: linux/x86_64
-    command: /app/celery/beat-start.sh
-    volumes:
-      - celery-data:/var/run/celery
-      - tusd-data:/srv/tusd-data/data
-      - user-data:/static
-    restart: always
-    depends_on:
-      - redis
-      - db
-    env_file:
-      - backend.env
-      - db.env
-    environment:
-      - TILE_SIGNING_SECRET_KEY=${TILE_SIGNING_SECRET_KEY}
-
-  flower:
[…]
