+++ docker-compose.prod.yml
@@ -1,429 +1,0 @@
-services:
-  frontend:
-    build:
-      args:
-        - VITE_STAC_ENABLED=true
-      context: ./frontend
-      dockerfile: frontend.prod.dockerfile
-    image: d2s-app:latest
-    platform: linux/x86_64
-    depends_on:
-      - backend
-    # read_only: true
-    tmpfs:
-      - /tmp
-    healthcheck:
-      test: ["CMD", "curl", "-f", "http://localhost:8080"]
-      interval: 1m30s
-      timeout: 30s
-      retries: 5
-      start_period: 30s
-    deploy:
-      restart_policy:
-        condition: on-failure
-        delay: 5s
-        max_attempts: 5
-        window: 120s
-    security_opt:
-      - no-new-privileges=true
-    env_file:
-      - frontend.env
-
-  backend:
-    build:
-      context: ./backend
-      dockerfile: backend.dockerfile
-      args:
-        NUM_OF_WORKERS: 1
-        LIMIT_MAX_REQUESTS: 10000
-    image: d2s-api:latest
-    platform: linux/x86_64
-    volumes:
-      - external-data:/srv/external-data
-      - log-data:/app/logs
-      - tusd-data:/srv/tusd-data/data
-      - user-data:/static
-    read_only: true
-    tmpfs:
-      - /var/tmp
-    depends_on:
-      db:
-        condition: service_healthy
-    healthcheck:
-      test: ["CMD", "curl", "-f", "http://localhost:5000/api/v1/health"]
-      interval: 1m30s
-      timeout: 30s
-      retries: 5
-      start_period: 30s
-    deploy:
-      restart_policy:
-        condition: on-failure
-        delay: 5s
-        max_attempts: 5
-        window: 120s
-    security_opt:
-      - no-new-privileges=true
-    environment:
-      - TILE_SIGNING_SECRET_KEY=${TILE_SIGNING_SECRET_KEY}
-      - OTEL_EXPORTER_OTLP_TRACES_ENDPOINT=http://otel-collector:4318
-    env_file:
-      - backend.env
-      - db.env
-
-  db:
-    image: postgis/postgis:15-3.3
-    platform: linux/x86_64
-    user: postgres
-    volumes:
-      - db-data:/var/lib/postgresql/data
-    read_only: true
-    tmpfs:
-      - /var/run/postgresql
-    healthcheck:
-      test:
-        [
-          "CMD-SHELL",
-          "pg_isready --username=$$POSTGRES_USER --dbname=$$POSTGRES_DB",
-        ]
-      interval: 10s
-      timeout: 5s
-      retries: 5
-      start_period: 30s
-    environment:
-      - POSTGRES_USER=d2s
-      - POSTGRES_DB=d2s
-      - PGTZ=Etc/UTC
-    deploy:
-      restart_policy:
-        condition: on-failure
-        delay: 5s
-        max_attempts: 5
-        window: 120s
-    security_opt:
-      - no-new-privileges=true
-    env_file:
-      - db.env
-
-  redis:
-    image: redis:alpine
-    platform: linux/x86_64
-    user: redis
-    read_only: true
-    healthcheck:
-      test: ["CMD", "redis-cli", "ping"]
-      interval: 1m30s
-      timeout: 30s
-      retries: 5
-      start_period: 30s
-    deploy:
-      restart_policy:
-        condition: on-failure
-        delay: 5s
-        max_attempts: 5
-        window: 120s
-    security_opt:
-      - no-new-privileges=true
-
-  celery_worker:
-    image: d2s-api:latest
-    platform: linux/x86_64
[…]
