+++ docker-compose.example.yml
@@ -182,7 +182,7 @@ services:
       - db.env
 
   titiler:
-    image: ghcr.io/developmentseed/titiler:0.19.0
+    image: ghcr.io/developmentseed/titiler:0.23.1
     volumes:
       - user-data:/static
     environment:
