### 2025-09-26 2000eda
add rotation symbol to bars with images
