### 2025-09-29 db1ee11
Update paper.md
Replace some words to clarify
