### 2025-10-01 fbeab9a
update readme with new screen captures, add tusd-data to gitignore, drop duplicate restart policy
