### 2025-10-01 962b90e
add ability to edit indoor project details and manage project member access
