+++ backend/app/tests/crud/test_indoor_project.py
@@ -1,11 +1,17 @@
 from datetime import datetime
 from typing import List
+from uuid import uuid4
 
+import pytest
 from sqlalchemy.orm import Session
 
 from app import crud
+from app.core.exceptions import PermissionDenied, ResourceNotFound
+from app.models.project_type import ProjectType
 from app.schemas.indoor_project import IndoorProjectCreate, IndoorProjectUpdate
+from app.schemas.role import Role
 from app.tests.utils.indoor_project import create_indoor_project
+from app.tests.utils.project_member import create_project_member
 from app.tests.utils.user import create_user
 
 
@@ -181,3 +187,320 @@ def test_deactivate_indoor_project(db: Session) -> None:
     assert deactivated_indoor_project.is_active is False
     assert deactivated_indoor_project.deactivated_at
     assert deactivated_indoor_project.deactivated_at < datetime.utcnow()
+
+
+def test_get_with_permission_owner_can_access_with_viewer_permission(
+    db: Session,
+) -> None:
+    """
+    Test that project owner can access project with VIEWER permission level.
+    """
+    # Create indoor project
+    indoor_project = create_indoor_project(db)
+
+    # Owner should be able to access with VIEWER permission
+    result = crud.indoor_project.get_with_permission(
+        db,
+        indoor_project_id=indoor_project.id,
+        user_id=indoor_project.owner_id,
+        required_permission=Role.VIEWER,
+    )
+
+    assert result is not None
+    assert result.id == indoor_project.id
+    assert getattr(result, "role", None) == Role.OWNER.value
+
+
+def test_get_with_permission_owner_can_access_with_manager_permission(
+    db: Session,
+) -> None:
+    """
+    Test that project owner can access project with MANAGER permission level.
+    """
+    # Create indoor project
+    indoor_project = create_indoor_project(db)
+
+    # Owner should be able to access with MANAGER permission
+    result = crud.indoor_project.get_with_permission(
+        db,
+        indoor_project_id=indoor_project.id,
+        user_id=indoor_project.owner_id,
+        required_permission=Role.MANAGER,
+    )
+
+    assert result is not None
+    assert result.id == indoor_project.id
+    assert getattr(result, "role", None) == Role.OWNER.value
+
+
+def test_get_with_permission_owner_can_access_with_owner_permission(
+    db: Session,
+) -> None:
+    """
+    Test that project owner can access project with OWNER permission level.
+    """
+    # Create indoor project
+    indoor_project = create_indoor_project(db)
+
+    # Owner should be able to access with OWNER permission
+    result = crud.indoor_project.get_with_permission(
+        db,
+        indoor_project_id=indoor_project.id,
+        user_id=indoor_project.owner_id,
+        required_permission=Role.OWNER,
+    )
+
+    assert result is not None
+    assert result.id == indoor_project.id
+    assert getattr(result, "role", None) == Role.OWNER.value
+
+
+def test_get_with_permission_manager_can_access_with_viewer_permission(
+    db: Session,
+) -> None:
+    """
+    Test that project manager can access project with VIEWER permission level.
+    """
+    # Create indoor project and manager user
+    indoor_project = create_indoor_project(db)
+    manager = create_user(db)
+
+    # Add manager as project member
+    create_project_member(
+        db,
+        role=Role.MANAGER,
+        member_id=manager.id,
+        project_uuid=indoor_project.id,
+        project_type=ProjectType.INDOOR_PROJECT,
+    )
+
+    # Manager should be able to access with VIEWER permission
+    result = crud.indoor_project.get_with_permission(
+        db,
+        indoor_project_id=indoor_project.id,
+        user_id=manager.id,
[…]
