### 2025-10-05 #8595 @jGaboardi
@minyj283

I've got 2 questions before I start looking for reviewers for your submission:

1. I am correct that users must [sign up for an account](https://docs.gdsl.org/data-to-science-user-manual/introduction-to-data-to-science/getting-started/sign-up) to use D2S?
2. If *\#1.* is "yes", is that account automatically approved from sign-up or is there a human "gatekeeper" that makes it a manual process?
