### 2025-10-06 #8595 @minyj283
Hello, @jGaboardi

1.	You are correct. Once a user sets up their own D2S instance, they need to sign up using their email address.
2.	By default, new accounts are automatically approved. However, administrators who deploy D2S on their computing resources can change this behavior through environment settings.

If these questions arise from concerns about the blind review process, please note that reviewers will run D2S on their local machines (e.g., at http://localhost:8000) using the D2S Docker image. Therefore, we will have no way of identifying who the reviewers are.

Alternatively, if the questions are about selecting appropriate reviewers, as described in the Quick Start section of our README, the installation of D2S is straightforward through the Docker image, and we believe anyone can easily test whether the system functions as intended.
