### 2025-10-06 #8595 @jGaboardi
@minyj283

Thanks for your quick response!

> If these questions arise from concerns about the blind review process
> f the questions are about selecting appropriate reviewers

These questions are regarding whether there is human in the loop to approve/verify when a user is trying to create an account. I think if the user *must* create the account to utilize D2S, that is still OK *if* that account approval process is entirely automated, similar to signing up for a GitHub account. I realize you have already [responded to this above](https://github.com/openjournals/joss-reviews/issues/8595#:~:text=Response%20to%20%22Illustrate%20what%20sorts%20of%20research%20this%20enables%20without%20a%20user%20needing%20to%20make%20a%20login%20at%20Purdue%22). However, we should be 100% clear on this point before proceeding.

cc @kthyng
