### 2025-10-06 #8595 @jGaboardi
> Please let us know if anything requires further clarification.

To be clear my actual question was _"Is there a human that decides on approval for account creation or is it automated?"_

However, after conferring with other editors it is probably OK either way, since the code itself is fully open.

I will begin attempting to find appropriate reviewers in the coming days.
