### 2025-10-07 #8595 @minyj283
@jGaboardi

1. Thank you. We will try to keep the word limits during the review process.
2. I just added the missing DOI in paper.lib.
