### 2025-10-15 b7c6d01
new endpoint for creating vector layers with geojson feature collections
