### 2025-10-16 4ccda27
optimize query for flights and data products
