### 2025-10-17 cc30173
cache bbox, crs, and resolution in data products table
