+++ backend/alembic/versions/13b80df8f03f_add_bbox_crs_resolution_to_data_products.py
@@ -1,0 +1,32 @@
+"""add bbox crs resolution to data products
+
+Revision ID: 13b80df8f03f
+Revises: 884e43c58053
+Create Date: 2025-10-16 16:40:47.348113
+
+"""
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision = '13b80df8f03f'
+down_revision: str | None = '884e43c58053'
+branch_labels: str | None = None
+depends_on: str | None = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('data_products', sa.Column('bbox', postgresql.JSONB(astext_type=sa.Text()), nullable=True))
+    op.add_column('data_products', sa.Column('crs', postgresql.JSONB(astext_type=sa.Text()), nullable=True))
+    op.add_column('data_products', sa.Column('resolution', postgresql.JSONB(astext_type=sa.Text()), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('data_products', 'resolution')
+    op.drop_column('data_products', 'crs')
+    op.drop_column('data_products', 'bbox')
+    # ### end Alembic commands ###
