+++ backend/app/crud/crud_data_product.py
@@ -14,6 +14,7 @@ from app import crud
 from app.api.utils import get_signature_for_data_product
 from app.core.config import settings
 from app.crud.base import CRUDBase
+from app.db.session import SessionLocal
 from app.models.constants import NON_RASTER_TYPES, PROCESSING_JOB_NAMES
 from app.models.data_product import DataProduct
 from app.models.job import Job
@@ -34,6 +35,15 @@ class CRUDDataProduct(CRUDBase[DataProduct, DataProductCreate, DataProductUpdate
     ) -> DataProduct:
         obj_in_data = jsonable_encoder(obj_in)
         data_product = DataProduct(**obj_in_data, flight_id=flight_id)
+
+        # Calculate and cache raster metadata if this is a raster product
+        if data_product.data_type not in NON_RASTER_TYPES:
+            metadata = calculate_raster_metadata(data_product.filepath)
+            if metadata:
+                data_product.bbox = metadata["bbox"]
+                data_product.crs = metadata["crs"]
+                data_product.resolution = metadata["resolution"]
+
         with db as session:
             session.add(data_product)
             session.commit()
@@ -57,7 +67,7 @@ class CRUDDataProduct(CRUDBase[DataProduct, DataProductCreate, DataProductUpdate
             data_product = session.execute(data_product_query).scalar_one_or_none()
             user_style = session.execute(user_style_query).scalar_one_or_none()
             if data_product:
-                set_bbox_attr(data_product)
+                set_spatial_metadata_attrs(data_product)
                 set_url_attr(data_product, upload_dir)
                 is_status_set = set_status_attr(data_product, data_product.jobs)
                 if user_style:
@@ -93,7 +103,7 @@ class CRUDDataProduct(CRUDBase[DataProduct, DataProductCreate, DataProductUpdate
         with db as session:
             data_product = session.scalar(data_product_query)
             if data_product and data_product.file_permission.is_public:
-                set_bbox_attr(data_product)
+                set_spatial_metadata_attrs(data_product)
                 set_signature_attr(data_product)
                 set_url_attr(data_product, upload_dir)
                 return data_product
@@ -103,7 +113,7 @@ class CRUDDataProduct(CRUDBase[DataProduct, DataProductCreate, DataProductUpdate
                     db, project_uuid=project_id, member_id=user_id
                 )
                 if project_member:
-                    set_bbox_attr(data_product)
+                    set_spatial_metadata_attrs(data_product)
                     set_signature_attr(data_product)
                     set_url_attr(data_product, upload_dir)
                     return data_product
@@ -137,7 +147,9 @@ class CRUDDataProduct(CRUDBase[DataProduct, DataProductCreate, DataProductUpdate
                 )
                 user_styles = session.execute(user_styles_query).scalars().all()
                 for user_style in user_styles:
-                    user_styles_by_data_product_id[user_style.data_product_id] = user_style
+                    user_styles_by_data_product_id[user_style.data_product_id] = (
+                        user_style
+                    )
 
             updated_data_products = []
             for data_product in data_products:
@@ -148,7 +160,7 @@ class CRUDDataProduct(CRUDBase[DataProduct, DataProductCreate, DataProductUpdate
                         set_user_style_attr(data_product, user_style.settings)
 
                 # Set additional attributes to be returned by API
-                set_bbox_attr(data_product)
+                set_spatial_metadata_attrs(data_product)
                 set_public_attr(data_product, data_product.file_permission.is_public)
                 set_signature_attr(data_product)
                 set_url_attr(data_product, upload_dir)
@@ -236,46 +248,66 @@ def set_status_attr(data_product_obj: DataProduct, jobs: List[Job]) -> bool:
         return True
 
 
-def set_bbox_attr(data_product: DataProduct) -> None:
[…]

[…]
