### 2025-10-17 4bc7622
display raster metadata from data product cards
