### 2025-10-20 ca7c324
add option to edit map layer names
