### 2025-10-22 c7ec3dd
replace gdalwarp with gdal_translate, calculate statistics during cog generation
