### 2025-10-22 46628f7
reduced database calls and improved filtering/aggregation when querying projects
