+++ backend/app/crud/crud_project.py
@@ -1,7 +1,6 @@
-from datetime import date
 import logging
 import json
-from typing import List, Optional, Sequence, Tuple, TypedDict
+from typing import List, Optional, Sequence, TypedDict
 from uuid import UUID
 
 from fastapi import status
@@ -22,7 +21,6 @@ from app.models.project_like import ProjectLike
 from app.models.project_member import ProjectMember
 from app.models.project_module import ProjectModule
 from app.models.project_type import ProjectType
-from app.models.team import Team
 from app.models.team_member import TeamMember
 from app.models.user import User
 from app.models.utils.utcnow import utcnow
@@ -33,7 +31,6 @@ from app.schemas.project import (
     Project as ProjectSchema,
     Projects,
 )
-from app.schemas.project_member import ProjectMemberCreate
 from app.schemas.team_member import Role
 
 
@@ -250,6 +247,35 @@ class CRUDProject(CRUDBase[Project, ProjectCreate, ProjectUpdate]):
         has_raster: bool = False,
         include_all: bool = False,
     ) -> List[Projects]:
+        # Build aggregation subquery for flight and data product counts
+        # This eliminates N+1 queries by computing counts in SQL
+        flight_stats_subquery = (
+            select(
+                Flight.project_id,
+                func.count(func.distinct(Flight.id))
+                .filter(Flight.is_active)
+                .label("flight_count"),
+                func.max(Flight.acquisition_date)
+                .filter(Flight.is_active)
+                .label("most_recent_flight"),
+                func.count(DataProduct.id)
+                .filter(and_(Flight.is_active, DataProduct.is_active))
+                .label("data_product_count"),
+                func.count(DataProduct.id)
+                .filter(
+                    and_(
+                        Flight.is_active,
+                        DataProduct.is_active,
+                        DataProduct.data_type != "point_cloud",
+                    )
+                )
+                .label("raster_count"),
+            )
+            .outerjoin(DataProduct, DataProduct.flight_id == Flight.id)
+            .group_by(Flight.project_id)
+            .subquery()
+        )
+
         # query to select active projects associated with user
         if include_all and user.is_superuser:
             statement = (
@@ -266,8 +292,22 @@ class CRUDProject(CRUDBase[Project, ProjectCreate, ProjectUpdate]):
                     )
                     .exists()
                     .label("liked"),
+                    func.coalesce(flight_stats_subquery.c.flight_count, 0).label(
+                        "flight_count"
+                    ),
+                    flight_stats_subquery.c.most_recent_flight,
+                    func.coalesce(flight_stats_subquery.c.data_product_count, 0).label(
+                        "data_product_count"
+                    ),
+                    func.coalesce(flight_stats_subquery.c.raster_count, 0).label(
+                        "raster_count"
+                    ),
                 )
                 .join(Project.location)
+                .outerjoin(
+                    flight_stats_subquery,
+                    flight_stats_subquery.c.project_id == Project.id,
+                )
                 .where(Project.is_active)
                 .options(selectinload(Project.team))
             )
@@ -287,28 +327,63 @@ class CRUDProject(CRUDBase[Project, ProjectCreate, ProjectUpdate]):
                     )
                     .exists()
                     .label("liked"),
+                    func.coalesce(flight_stats_subquery.c.flight_count, 0).label(
+                        "flight_count"
+                    ),
+                    flight_stats_subquery.c.most_recent_flight,
+                    func.coalesce(flight_stats_subquery.c.data_product_count, 0).label(
+                        "data_product_count"
+                    ),
+                    func.coalesce(flight_stats_subquery.c.raster_count, 0).label(
+                        "raster_count"
+                    ),
                 )
                 .join(Project.members)
[…]
