### 2025-10-22 dde883f
check if band stats are missing and compute if so
