### 2025-10-24 e31ed30
refactor stac related classes and utility functions
