+++ backend/app/utils/STACCollectionManager.py
@@ -1,408 +1,0 @@
-import logging
-import os
-from datetime import datetime, timezone
-from typing import Dict, List, Optional
-
-import requests
-from fastapi.encoders import jsonable_encoder
-from pystac import Collection, Item
-from pystac_client import Client
-from pystac_client.exceptions import APIError
-
-from app.core.config import settings
-from app.schemas import STACReport, ItemStatus, STACError
-
-logger = logging.getLogger(__name__)
-
-
-class STACCollectionManager:
-    @property
-    def STAC_API_URL(self) -> Optional[str]:
-        return str(settings.get_stac_api_url) if settings.get_stac_api_url else None
-
-    def __init__(
-        self,
-        collection_id: str,
-        collection: Optional[Collection] = None,
-        items: Optional[List[Item]] = None,
-    ) -> None:
-        if not self.STAC_API_URL:
-            logger.error("STAC_API_URL environment setting is not set")
-            raise ValueError("STAC_API_URL environment setting is not set.")
-
-        if not self.__class__.is_stac_api_available():
-            logger.error(f"STAC API at {self.STAC_API_URL} is unavailable")
-            raise ConnectionError(f"STAC API at {self.STAC_API_URL} is unavailable.")
-
-        self.api_url = self.STAC_API_URL
-        self.collection_id = collection_id
-        self.collection = collection
-        self.items = items or []
-
-        logger.debug(
-            f"Initialized STACCollectionManager for collection {collection_id}"
-        )
-
-    def fetch_public_collection(self) -> Optional[Dict]:
-        """Get STAC Collection in public catalog.
-
-        Raises:
-            APIError: Raised if client fails to connect due to API being down.
-            ValueError: Raised if collection cannot be transformed to dict.
-
-        Returns:
-            Optional[Dict]: Collection as dict, or None if collection not found.
-        """
-        try:
-            # Connect to STAC API and retrieve collection
-            logger.info(f"Connecting to STAC API at {self.api_url}")
-            client = Client.open(self.api_url)
-            logger.debug(f"Fetching collection {self.collection_id}")
-            collection = client.get_collection(self.collection_id)
-        except APIError as e:
-            # Check if this is a 404 error (collection not found)
-            if hasattr(e, "status_code") and e.status_code == 404:
-                logger.info(
-                    f"Collection {self.collection_id} not found in STAC catalog"
-                )
-                return None
-            else:
-                # This is a different API error (like connection issue)
-                logger.error(
-                    f"Failed to connect to STAC API or fetch collection {self.collection_id}: {str(e)}"
-                )
-                raise
-
-        if not collection:
-            logger.info(f"Collection {self.collection_id} not found in STAC catalog")
-            return None
-
-        try:
-            # Transform Collection to dict
-            logger.debug(f"Transforming collection {self.collection_id} to dict")
-            collection_d = collection.to_dict()
-        except Exception as e:
-            logger.error(
-                f"Failed to transform collection {self.collection_id} to dict: {str(e)}"
-            )
-            raise ValueError(f"Failed to transform collection to dict: {str(e)}")
-
-        logger.info(f"Successfully fetched collection {self.collection_id}")
-        return collection_d
-
-    def fetch_public_items(self) -> List[str]:
-        """Get STAC Items in public catalog.
-
-        Returns:
-            List[str]: List of published item ids associated with collection.
-        """
-        try:
-            # Connect to STAC API and retrieve items
-            logger.info(f"Connecting to STAC API at {self.api_url}")
-            client = Client.open(self.api_url)
-            logger.debug(f"Fetching items for collection {self.collection_id}")
-            items = client.get_collection(self.collection_id).get_items()
-
-            # Get item ids from items
[…]
