### 2025-10-27 acdfaa7
allow user to include raw data links for select data products
