### 2025-10-28 544e34a
better logging and use of stac cache
