+++ backend/app/crud/crud_raw_data.py
@@ -1,3 +1,4 @@
+import logging
 import os
 from pathlib import Path
 from typing import List, Sequence
@@ -14,6 +15,8 @@ from app.models.job import Job
 from app.models.raw_data import RawData
 from app.models.utils.utcnow import utcnow
 from app.schemas.raw_data import RawDataCreate, RawDataUpdate
+
+logger = logging.getLogger(__name__)
 
 
 class CRUDRawData(CRUDBase[RawData, RawDataCreate, RawDataUpdate]):
@@ -128,7 +131,11 @@ def set_url_attr(raw_data_obj: RawData, upload_dir: str):
         static_url = settings.API_DOMAIN + settings.STATIC_DIR
         relative_path = Path(raw_data_obj.filepath).relative_to(upload_dir)
         setattr(raw_data_obj, "url", f"{static_url}/{str(relative_path)}")
-    except ValueError:
+    except ValueError as e:
+        logger.warning(
+            f"Unable to compute URL for raw data {raw_data_obj.id}: "
+            f"filepath '{raw_data_obj.filepath}' is not relative to upload_dir '{upload_dir}'. {e}"
+        )
         setattr(raw_data_obj, "url", None)
 
 
