### 2025-10-28 5cd3473
add file permission controls for raw data
