### 2025-11-03 5be5083
add control for adjusting point size in side menu
