### 2025-11-05 198c049
add contacts extension and contact form on stac management page
