### 2025-11-05 ccc1a59
add link for external viewer application to published stac items
