### 2025-11-09 #8595 @jGaboardi
> Sorry, but I am one of the co-authors of the paper.

@zivankaraman – Please clarify this statement. @minyj283 has not listed you as a co-author and I can't find you in the [`data-to-science`](https://github.com/gdslab/data-to-science) commit history.

cc @kthyng
