### 2025-11-10 c6fa05f
task and utility methods for generating geoparquet files, backfill script, automated fallback on static file request
