+++ backend/app/api/utils.py
@@ -6,6 +6,7 @@ from pathlib import Path
 from typing import List, Optional, Tuple
 from urllib.parse import urlencode, quote_plus
 
+import geopandas as gpd
 from geojson_pydantic import Feature
 from pydantic import UUID4
 from sqlalchemy.orm import Session
@@ -74,6 +75,35 @@ def create_vector_layer_preview(
         vector_layer_preview.save(outname="preview.png")
 
     return preview_img
+
+
+def save_vector_layer_parquet(
+    project_id: uuid.UUID, layer_id: str, gdf: gpd.GeoDataFrame, static_dir: str
+) -> str:
+    """Generate and save GeoParquet file for a vector layer.
+
+    Args:
+        project_id (uuid.UUID): Project ID.
+        layer_id (str): Unique layer ID for FeatureCollection.
+        gdf (gpd.GeoDataFrame): GeoDataFrame containing vector layer features.
+        static_dir (str): Path to static directory (from get_static_dir()).
+
+    Returns:
+        str: Path to generated parquet file.
+    """
+    # Set output path for parquet file
+    parquet_dir = os.path.join(
+        static_dir, "projects", str(project_id), "vector", layer_id
+    )
+    # Create vector directory if needed
+    if not os.path.exists(parquet_dir):
+        os.makedirs(parquet_dir)
+    # Full path to parquet file
+    parquet_path = os.path.join(parquet_dir, f"{layer_id}.parquet")
+    # Save GeoDataFrame as GeoParquet with snappy compression
+    gdf.to_parquet(parquet_path, compression="snappy")
+
+    return parquet_path
 
 
 def is_valid_uuid(id: str) -> bool:
