+++ backend/app/tests/utils/test_utils.py
@@ -1,14 +1,20 @@
 import os
 from datetime import datetime, timedelta, timezone
 from typing import List
+from uuid import UUID
 
 import geopandas as gpd
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
 from app import crud
-from app.api.utils import create_vector_layer_preview
+from app.api.utils import (
+    create_vector_layer_preview,
+    save_vector_layer_parquet,
+    get_static_dir,
+)
 from app.core.config import settings
+from app.utils.ProtectedStaticFiles import parse_vector_parquet_path
 from app.models.data_product import DataProduct
 from app.models.flight import Flight
 from app.models.job import Job
@@ -79,6 +85,135 @@ def test_create_vector_layer_preview_image(
     assert os.path.exists(point_preview)
     assert os.path.exists(line_preview)
     assert os.path.exists(polygon_preview)
+
+
+def test_save_vector_layer_parquet(db: Session) -> None:
+    """Test that save_vector_layer_parquet creates valid parquet files."""
+    # Create project
+    project = create_project(db)
+
+    # Test with Point geometry
+    point_vector_layer: VectorLayerDict = get_geojson_feature_collection("point")
+    gdf_point = gpd.GeoDataFrame.from_features(
+        point_vector_layer["geojson"]["features"], crs="EPSG:4326"
+    )
+    point_features = crud.vector_layer.create_with_project(
+        db, file_name=point_vector_layer["layer_name"], gdf=gdf_point, project_id=project.id
+    )
+    point_layer_id = point_features[0].properties["layer_id"]
+
+    # Generate parquet file
+    static_dir = get_static_dir()
+    parquet_path = save_vector_layer_parquet(
+        project_id=project.id,
+        layer_id=point_layer_id,
+        gdf=gdf_point,
+        static_dir=static_dir,
+    )
+
+    # Verify parquet file exists
+    assert os.path.exists(parquet_path)
+    expected_path = os.path.join(
+        settings.TEST_STATIC_DIR,
+        "projects",
+        str(project.id),
+        "vector",
+        point_layer_id,
+        f"{point_layer_id}.parquet",
+    )
+    assert parquet_path == expected_path
+
+    # Verify parquet file can be read
+    gdf_from_parquet = gpd.read_parquet(parquet_path)
+    assert len(gdf_from_parquet) == len(gdf_point)
+    assert gdf_from_parquet.crs == gdf_point.crs
+
+    # Test with Polygon geometry
+    polygon_vector_layer: VectorLayerDict = get_geojson_feature_collection("polygon")
+    gdf_polygon = gpd.GeoDataFrame.from_features(
+        polygon_vector_layer["geojson"]["features"], crs="EPSG:4326"
+    )
+    polygon_features = crud.vector_layer.create_with_project(
+        db, file_name=polygon_vector_layer["layer_name"], gdf=gdf_polygon, project_id=project.id
+    )
+    polygon_layer_id = polygon_features[0].properties["layer_id"]
+
+    # Generate parquet file for polygon
+    parquet_path_polygon = save_vector_layer_parquet(
+        project_id=project.id,
+        layer_id=polygon_layer_id,
+        gdf=gdf_polygon,
+        static_dir=static_dir,
+    )
+
+    # Verify polygon parquet file
+    assert os.path.exists(parquet_path_polygon)
+    gdf_polygon_from_parquet = gpd.read_parquet(parquet_path_polygon)
+    assert len(gdf_polygon_from_parquet) == len(gdf_polygon)
+
+    # Test with LineString geometry
[…]
