### 2025-11-11 684d4b8
map layers table layout improvements, more consistent download names
