### 2025-11-11 0212032
generate FlatGeobuf file for vector layers
