+++ backend/app/utils/ProtectedStaticFiles.py
@@ -15,7 +15,11 @@ from app.utils.staticfiles import RangedStaticFiles
 
 from app import crud
 from app.api.deps import can_read_project
-from app.api.utils import save_vector_layer_parquet, get_static_dir
+from app.api.utils import (
+    save_vector_layer_parquet,
+    save_vector_layer_flatgeobuf,
+    get_static_dir,
+)
 from app.core import security
 from app.db.session import SessionLocal
 from app.models.data_product import DataProduct
@@ -36,6 +40,32 @@ def parse_vector_parquet_path(path: str) -> Optional[Tuple[UUID, str]]:
     """
     # Pattern to match: /static/projects/{uuid}/vector/{layer_id}/{layer_id}.parquet
     pattern = r"/static/projects/([0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12})/vector/([^/]+)/\2\.parquet"
+    match = re.match(pattern, path, re.IGNORECASE)
+
+    if match:
+        try:
+            project_id = UUID(match.group(1))
+            layer_id = match.group(2)
+            return (project_id, layer_id)
+        except ValueError:
+            return None
+
+    return None
+
+
+def parse_vector_flatgeobuf_path(path: str) -> Optional[Tuple[UUID, str]]:
+    """Parse vector FlatGeobuf path to extract project_id and layer_id.
+
+    Expected path format: /static/projects/{uuid}/vector/{layer_id}/{layer_id}.fgb
+
+    Args:
+        path: URL path to parse
+
+    Returns:
+        Tuple of (project_id, layer_id) if valid FlatGeobuf path, None otherwise
+    """
+    # Pattern to match: /static/projects/{uuid}/vector/{layer_id}/{layer_id}.fgb
+    pattern = r"/static/projects/([0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12})/vector/([^/]+)/\2\.fgb"
     match = re.match(pattern, path, re.IGNORECASE)
 
     if match:
@@ -238,28 +268,7 @@ async def verify_static_file_access(request: Request) -> None:
         if file_permission and file_permission.is_public:
             return
 
-    # restricted access authorization
-    access_token = request.cookies.get("access_token")
-    if not access_token:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
-
-    # Extract token from "Bearer <token>" format
-    token = access_token.split(" ")[1]
-
-    # Validate token and get payload
-    payload = security.validate_token_and_get_payload(token, "access")
-
-    # Get user from payload
-    db = SessionLocal()
-    try:
-        user = security.get_user_from_token_payload(db, payload)
-    finally:
-        db.close()
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="user not found"
-        )
+    # Handle project-scoped files (vector data, previews, etc.)
     if "projects" in request.url.path:
         try:
             project_id = request.url.path.split("/projects/")[1].split("/")[0]
@@ -268,6 +277,49 @@ async def verify_static_file_access(request: Request) -> None:
             raise HTTPException(
                 status_code=status.HTTP_404_NOT_FOUND, detail="project not found"
             )
+
+        # Check for API_KEY authentication (for programmatic access, QGIS, etc.)
+        if "API_KEY" in request.query_params:
+            api_key = request.query_params["API_KEY"]
+            db = SessionLocal()
+            try:
+                api_key_obj = crud.api_key.get_by_api_key(db, api_key)
+                if api_key_obj:
+                    user_from_api_key = api_key_obj.owner
[…]
