### 2025-11-12 97a281d
improved validation and API key handling for static files
