### 2025-11-12 4f3ce00
improve email messaging and branding consistency
