### 2025-11-13 #8595 @zivankaraman
_> So, to answer your initial question, I will have a look at the submitted paper and will let you know shortly._

I read the submitted paper about D2S software platform for geospatial data science research. This is a very interesting and valuable paper and software. However, it is based on software tools that I am not sufficiently familiar with, which prevents me from adequately reviewing this paper.
