### 2025-11-13 f45b639
refine attributes returned by users endpoint, create new endpoint for admins
