### 2025-11-14 9741fd8
add vector layer static file urls to payload, fix mobile layout issues with map layers table
