### 2025-11-16 f97a9b0
include count of public data products
