### 2025-11-16 2532233
add last_login_at and last_activity_at to users, middleware to update last activity
