+++ backend/app/main.py
@@ -1,19 +1,17 @@
 import logging
-import time
-from typing import Awaitable, Callable
 
-from fastapi import FastAPI, HTTPException, Request, Response
+from fastapi import FastAPI, HTTPException, Request
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse, RedirectResponse
 from fastapi.staticfiles import StaticFiles
 from opentelemetry.instrumentation.asgi import OpenTelemetryMiddleware
 from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
-from starlette.background import BackgroundTask
 
 from app.api.api_v1.api import api_router
 from app.api.extras import extra_router
 from app.core.config import settings
-from app.core.logging import get_http_info, setup_logger
+from app.core.logging import setup_logger
+from app.middleware import log_http_request, track_user_activity
 from app.utils.ProtectedStaticFiles import ProtectedStaticFiles
 
 
@@ -72,25 +70,6 @@ def http_exception_handler(request: Request, exc: HTTPException) -> JSONResponse
     return JSONResponse(status_code=exc.status_code, content={"detail": exc.detail})
 
 
-def write_access_log(request: Request, response: Response, process_time: float) -> None:
-    http_info = get_http_info(request, response)
-    http_info["res"]["process_time"] = f"{process_time:.3f}"
-    logger.info(
-        request.method + " " + request.url.path,
-        extra={"extra_info": http_info},
-    )
-
-
-@app.middleware("http")
-async def log_http_request(
-    request: Request, call_next: Callable[[Request], Awaitable[Response]]
-) -> Response:
-    start_time = time.time()
-    response = await call_next(request)
-    process_time = time.time() - start_time
-
-    response.background = BackgroundTask(
-        write_access_log, request, response, process_time
-    )
-
-    return response
+# Register middleware
+app.middleware("http")(log_http_request)
+app.middleware("http")(track_user_activity)
