+++ backend/alembic/versions/b65c59b096f4_add_last_login_at_and_last_activity_at_.py
@@ -1,0 +1,34 @@
+"""add last_login_at and last_activity_at to users table
+
+Revision ID: b65c59b096f4
+Revises: c186d6643f9f
+Create Date: 2025-11-16 15:29:47.441757
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'b65c59b096f4'
+down_revision: str | None = 'c186d6643f9f'
+branch_labels: str | None = None
+depends_on: str | None = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('users', sa.Column('last_login_at', sa.DateTime(timezone=True), nullable=True))
+    op.add_column('users', sa.Column('last_activity_at', sa.DateTime(timezone=True), nullable=True))
+    op.create_index(op.f('ix_users_last_activity_at'), 'users', ['last_activity_at'], unique=False)
+    op.create_index(op.f('ix_users_last_login_at'), 'users', ['last_login_at'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_users_last_login_at'), table_name='users')
+    op.drop_index(op.f('ix_users_last_activity_at'), table_name='users')
+    op.drop_column('users', 'last_activity_at')
+    op.drop_column('users', 'last_login_at')
+    # ### end Alembic commands ###
