### 2025-11-17 3d5b585
update quickstart image references, handle empty URL env vars
