### 2025-11-18 01bf4f2
Update README with HTTP_COOKIE_SECURE instructions
Added instructions for configuring HTTP_COOKIE_SECURE for non-localhost access.
