### 2025-11-19 8faa423
handle multiple middleware background tasks and improve logging
