### 2025-11-22 #8595 @jGaboardi
@morrowcj

> I'm not very familiar with JavaScript, but if that's OK, then I'd be willing to serve as a reviewer.

Yes, I believe that is perfectly acceptable, because there is also the Python portion.

Thanks! I'll sign you up and we can start the review!
