### 2025-11-22 #9497 @editorialbot
[REVIEW]: Data-to-Science (D2S): An open-source ecosystem for collaborative geospatial data science research
**Submitting author:** @minyj283 (<a href="http://orcid.org/0000-0002-6419-5236">Minyoung Jung</a>)
**Repository:** https://github.com/gdslab/data-to-science
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
