### 2025-12-01 c807562
make clear distinction between two project interfaces, update imports
