### 2025-12-03 e971549
upgrade axios to new minor version, upgrade react router from v6 to v7
