### 2025-12-04 abeac26
upgrade eslint and fix linter errors
