+++ frontend/src/components/pages/projects/stac/hooks/useSTACActions.ts
@@ -29,6 +29,32 @@ interface UseSTACActionsReturn {
   handleUnpublish: () => Promise<void>;
 }
 
+const actionConfigs: Record<ActionType, ActionConfig> = {
+  publish: {
+    method: 'PUT',
+    endpoint: (id) => `/projects/${id}/publish-stac-async`,
+    successMessage:
+      'Publish request sent. Please check back in a few minutes.',
+    errorMessage: 'Failed to publish project',
+    operationType: 'publishing',
+  },
+  update: {
+    method: 'PUT',
+    endpoint: (id) => `/projects/${id}/publish-stac-async`,
+    successMessage:
+      'Update request sent. Please check back in a few minutes.',
+    errorMessage: 'Failed to update STAC catalog',
+    operationType: 'updating',
+  },
+  unpublish: {
+    method: 'DELETE',
+    endpoint: (id) => `/projects/${id}/delete-stac`,
+    successMessage: 'Project unpublished successfully',
+    errorMessage: 'Failed to unpublish project',
+    operationType: 'unpublishing',
+  },
+};
+
 export function useSTACActions({
   projectId,
   buildRequestPayload,
@@ -36,32 +62,6 @@ export function useSTACActions({
   setStatus,
 }: UseSTACActionsProps): UseSTACActionsReturn {
   const revalidator = useRevalidator();
-
-  const actionConfigs: Record<ActionType, ActionConfig> = {
-    publish: {
-      method: 'PUT',
-      endpoint: (id) => `/projects/${id}/publish-stac-async`,
-      successMessage:
-        'Publish request sent. Please check back in a few minutes.',
-      errorMessage: 'Failed to publish project',
-      operationType: 'publishing',
-    },
-    update: {
-      method: 'PUT',
-      endpoint: (id) => `/projects/${id}/publish-stac-async`,
-      successMessage:
-        'Update request sent. Please check back in a few minutes.',
-      errorMessage: 'Failed to update STAC catalog',
-      operationType: 'updating',
-    },
-    unpublish: {
-      method: 'DELETE',
-      endpoint: (id) => `/projects/${id}/delete-stac`,
-      successMessage: 'Project unpublished successfully',
-      errorMessage: 'Failed to unpublish project',
-      operationType: 'unpublishing',
-    },
-  };
 
   const executeAction = useCallback(
     async (actionType: ActionType) => {
@@ -86,7 +86,7 @@ export function useSTACActions({
           msg: config.successMessage,
         });
         revalidator.revalidate();
-      } catch (err) {
+      } catch {
         setStatus({
           type: 'error',
           msg: config.errorMessage,
@@ -95,14 +95,7 @@ export function useSTACActions({
         setCurrentOperation('idle');
       }
     },
-    [
-      projectId,
-      buildRequestPayload,
-      setCurrentOperation,
-      setStatus,
-      revalidator,
-      actionConfigs,
-    ]
+    [projectId, buildRequestPayload, setCurrentOperation, setStatus, revalidator]
   );
 
   return {
