+++ frontend/src/types/pannellum.d.ts
@@ -1,0 +1,16 @@
+declare global {
+  interface Window {
+    pannellum: {
+      viewer: (
+        container: HTMLElement | null,
+        config: {
+          type: string;
+          panorama: string;
+          autoLoad: boolean;
+        }
+      ) => void;
+    };
+  }
+}
+
+export {};
