+++ frontend/src/types/uppy.d.ts
@@ -1,0 +1,11 @@
+// Types for Uppy upload error handling
+
+export type ValidationError = {
+  loc: (string | number)[];
+  msg: string;
+  type: string;
+};
+
+export type ErrorResponseBody = {
+  detail?: string | ValidationError[];
+};
