### 2025-12-12 gdslab/data-to-science#299 @morrowcj
JOSS review: Add author contributions to paper.md
I recommend adding an "Author Contributions" section to docs/paper.md outlining the roles that each author contributed to the software. The [Contributor Role Taxonomy](https://credit.niso.org/) framework is preferred option.

As per the Joss Review Checklist (https://github.com/openjournals/joss-reviews/issues/9497):

> Contribution and authorship: Has the submitting author (@minyj283) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

In terms of commits, the submitting author appears to have made only 2 (out of >1100 total), containing only 9 lines of changes ([data-to-science/graphs/contributors](https://github.com/gdslab/data-to-science/graphs/contributors)). Their contributions, and those of all authors, should be made clear.
