### 2025-12-15 9f13f26
replace uppy drag and drop with dashboard modal
