+++ backend/app/tests/api/api_v1/test_users.py
@@ -1,5 +1,6 @@
 import os
 from typing import List
+from unittest.mock import AsyncMock, patch
 
 from fastapi import Request, status
 from fastapi.testclient import TestClient
@@ -22,6 +23,11 @@ from app.tests.utils.user import create_user
 from app.tests.conftest import pytest_requires_mail
 from app.tests.utils.utils import random_email, random_full_name, random_password
 
+# Turnstile test tokens
+TURNSTILE_SUCCESS_TOKEN = "1x0000000000000000000000000000000AA"
+TURNSTILE_FAILURE_TOKEN_1 = "2x0000000000000000000000000000000AA"
+TURNSTILE_FAILURE_TOKEN_2 = "3x0000000000000000000000000000000AA"
+
 
 @pytest_requires_mail
 def test_create_user_new_email(client: TestClient, db: Session) -> None:
@@ -38,7 +44,8 @@ def test_create_user_new_email(client: TestClient, db: Session) -> None:
     if fm:
         fm.config.SUPPRESS_SEND = 1
         with fm.record_messages() as outbox:
-            r = client.post(f"{settings.API_V1_STR}/users/", json=data)
+            with patch.object(settings, "TURNSTILE_SECRET_KEY", None):
+                r = client.post(f"{settings.API_V1_STR}/users/", json=data)
         assert 201 == r.status_code
         created_user = r.json()
         user = crud.user.get_by_email(db, email=data["email"])
@@ -50,13 +57,19 @@ def test_create_user_new_email(client: TestClient, db: Session) -> None:
             == settings.MAIL_FROM_NAME + " <" + settings.MAIL_FROM + ">"
         )
         assert outbox[0]["To"] == user.email
-        assert outbox[0]["Subject"] == "Welcome to Data to Science - please confirm your email"
+        assert (
+            outbox[0]["Subject"]
+            == "Welcome to Data to Science - please confirm your email"
+        )
         assert (
             outbox[1]["from"]
             == settings.MAIL_FROM_NAME + " <" + settings.MAIL_FROM + ">"
         )
         assert outbox[1]["To"] == settings.MAIL_ADMINS.replace(",", ", ")
-        assert outbox[1]["Subject"] == f"New Data to Science account awaiting approval ({api_domain})"
+        assert (
+            outbox[1]["Subject"]
+            == f"New Data to Science account awaiting approval ({api_domain})"
+        )
 
 
 def test_create_user_existing_email(client: TestClient, db: Session) -> None:
@@ -68,7 +81,8 @@ def test_create_user_existing_email(client: TestClient, db: Session) -> None:
         "first_name": existing_user.first_name,
         "last_name": existing_user.last_name,
     }
-    r = client.post(f"{settings.API_V1_STR}/users/", json=data)
+    with patch.object(settings, "TURNSTILE_SECRET_KEY", None):
+        r = client.post(f"{settings.API_V1_STR}/users/", json=data)
     assert 409 == r.status_code
 
 
@@ -81,7 +95,8 @@ def test_create_user_is_not_approved(client: TestClient, db: Session) -> None:
         "first_name": full_name["first"],
         "last_name": full_name["last"],
     }
-    r = client.post(f"{settings.API_V1_STR}/users/", json=data)
+    with patch.object(settings, "TURNSTILE_SECRET_KEY", None):
+        r = client.post(f"{settings.API_V1_STR}/users/", json=data)
     assert 201 == r.status_code
     created_user = r.json()
     if settings.MAIL_ENABLED:
@@ -100,8 +115,9 @@ def test_create_user_with_password_less_than_minimum_length(
         "first_name": full_name["first"],
         "last_name": full_name["last"],
     }
[…]
