### 2025-12-16 6d14526
add registration intent form field
