+++ backend/app/models/user.py
@@ -39,6 +39,9 @@ class User(Base):
     hashed_password: Mapped[str] = mapped_column(String(60), nullable=False)
     first_name: Mapped[str] = mapped_column(String(64), nullable=False)
     last_name: Mapped[str] = mapped_column(String(64), nullable=False)
+    registration_intent: Mapped[Optional[str]] = mapped_column(
+        String(500), nullable=True
+    )
     full_name: Mapped[str] = column_property(first_name + " " + last_name)
     is_approved: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
     is_demo: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
@@ -92,6 +95,7 @@ class User(Base):
         return (
             f"User(id={self.id!r}, email={self.email!r}, "
             f"first_name={self.first_name!r}, last_name={self.last_name!r}, "
+            f"registration_intent={self.registration_intent!r}, "
             f"is_approved={self.is_approved!r}, is_demo={self.is_demo!r}, "
             f"is_email_confirmed={self.is_email_confirmed!r}, "
             f"is_superuser={self.is_superuser!r}, created_at={self.created_at!r}, "
