+++ backend/alembic/versions/b59fcf1c51a7_add_registration_intent_to_users_table.py
@@ -1,0 +1,28 @@
+"""add registration_intent to users table
+
+Revision ID: b59fcf1c51a7
+Revises: b65c59b096f4
+Create Date: 2025-12-16 20:28:02.563427
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'b59fcf1c51a7'
+down_revision: str | None = 'b65c59b096f4'
+branch_labels: str | None = None
+depends_on: str | None = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('users', sa.Column('registration_intent', sa.String(length=500), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('users', 'registration_intent')
+    # ### end Alembic commands ###
