### 2025-12-17 gdslab/data-to-science#305 @morrowcj
JOSS Review: Documentation
Below are some suggested revisions to your documentation, as part of [JOSS REVIEW 9497](https://github.com/openjournals/joss-reviews/issues/9497)

## Community Guidelines

* Add a [CONTRIBUTING.md](https://contributing.md/example/) document to your repository to outline where users should report bugs,  suggest features, etc.

## Example Usage

* The paper.md states in the table in the "Collaboration" row that the software enables "Granting public access to data with no account or API key". It would be great if an example of this feature were present in the README (e.g., a link to some public data hosted on the platform), so that potential users can see this feature before committing to creating an account.

* Example usage are present in the user manuals, but not in the README. Further, the manuals are not linked in the README (they are linked in paper.md, which *is* linked in the README). Their inclusion would be appreciated for users.

* A link to the D2S website would also be nice in the README.

## User Manual (https://docs.gdsl.org/data-to-science-user-manual)

* The link on your [Getting Started](https://docs.gdsl.org/data-to-science-user-manual/introduction-to-data-to-science/getting-started) page is broken and goes nowhere.
