### 2026-01-05 cd0d997
add contributing guidelines and examples of public data access
