+++ CONTRIBUTING.md
@@ -1,0 +1,35 @@
+# Contributing to Data to Science (D2S)
+
+We welcome contributions, feedback, and discussion from the community.
+
+## Reporting Bugs
+
+If you encounter a bug or unexpected behavior, please open an issue on GitHub:
+
+- https://github.com/gdslab/data-to-science/issues
+
+When possible, include:
+
+- A brief description of the problem
+- Steps to reproduce the issue
+- Expected and observed behavior
+- Relevant logs, screenshots, or configuration details
+
+## Suggesting Features and Enhancements
+
+Feature requests and ideas for improvements are welcome. Please start a discussion or open an issue on GitHub:
+
+- Discussions: https://github.com/gdslab/data-to-science/discussions
+- Issues: https://github.com/gdslab/data-to-science/issues
+
+Discussions are recommended for exploratory ideas or questions, while issues are best for concrete, actionable feature requests.
+
+## Asking Questions and Starting Conversations
+
+General questions, use-case discussions, and design conversations can be started in GitHub Discussions:
+
+- https://github.com/gdslab/data-to-science/discussions
+
+## Contributing Code
+
+We welcome code contributions. If you plan to make a significant change, we recommend opening an issue or discussion first to coordinate and avoid duplicated effort.
