### 2026-01-05 26d4bd7
resource links for user manual and website/tutorial videos
