+++ docs/paper.bib
@@ -53,7 +53,8 @@
   number={2},
   pages={95--109},
   year={2023},
-  publisher={MDPI}
+  publisher={MDPI},
+  doi={10.3390/standards3020009}
 }
 
 @article{wu2020geemap,
