### 2026-01-07 4edb1b2
drop start_date and end_date from indoor project form, automatically populate with spreadsheet
