+++ backend/app/schemas/indoor_project.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from enum import Enum
-from typing import Any, Dict, Literal, Optional
+from typing import Any, Dict, Literal, Optional, Union
 
 from pydantic import BaseModel, Field, field_validator, ValidationInfo, UUID4
 
@@ -21,6 +21,14 @@ class IndoorProjectBase(BaseModel):
     start_date: Optional[datetime] = None
     end_date: Optional[datetime] = None
     team_id: Optional[UUID4] = None
+
+    @field_validator("start_date", "end_date", mode="before")
+    @classmethod
+    def empty_string_to_none(cls, v: Union[str, datetime, None]) -> Union[datetime, None]:
+        """Convert empty strings to None to handle form inputs."""
+        if v == "" or v is None:
+            return None
+        return v
 
     @field_validator("end_date")
     @classmethod
