+++ backend/app/tests/tasks/test_upload_tasks.py
@@ -1,0 +1,462 @@
+"""Tests for upload tasks, specifically date extraction from indoor project spreadsheets."""
+
+import pytest
+from datetime import datetime
+from uuid import uuid4
+from unittest.mock import patch
+import pandas as pd
+
+from app.tasks.upload_tasks import extract_dates_from_indoor_spreadsheet
+
+
+@patch("app.tasks.upload_tasks.pd.read_excel")
+def test_extract_dates_happy_path(mock_read_excel):
+    """Test successful date extraction with both PPEW and Top worksheets."""
+    test_id = uuid4()
+
+    # Mock PPEW DataFrame with planting date
+    mock_ppew_df = pd.DataFrame({
+        "VARIETY": ["Test Variety"],
+        "PI": ["12345"],
+        "PLANTING DATE": ["2024-01-15"],
+    })
+
+    # Mock Top DataFrame with scan dates
+    mock_top_df = pd.DataFrame({
+        "VARIETY": ["Test Variety", "Test Variety", "Test Variety"],
+        "SCAN DATE": ["2024-02-01", "2024-02-15", "2024-03-01"],
+    })
+
+    # Set up the mock to return different DataFrames based on sheet_name
+    def read_excel_side_effect(file_path, sheet_name, **kwargs):
+        if sheet_name == "PPEW":
+            return mock_ppew_df
+        elif sheet_name == "Top":
+            return mock_top_df
+        raise ValueError(f"Worksheet {sheet_name} not found")
+
+    mock_read_excel.side_effect = read_excel_side_effect
+
+    # Call the function
+    start_date, end_date = extract_dates_from_indoor_spreadsheet(
+        "/fake/path/test.xlsx", test_id
+    )
+
+    # Assertions
+    assert start_date is not None
+    assert end_date is not None
+    assert start_date.year == 2024
+    assert start_date.month == 1
+    assert start_date.day == 15
+    assert end_date.year == 2024
+    assert end_date.month == 3
+    assert end_date.day == 1
+
+
+@patch("app.tasks.upload_tasks.pd.read_excel")
+def test_extract_dates_missing_ppew_uses_scan_date_fallback(mock_read_excel):
+    """Test that min scan_date is used for start_date when PPEW is missing."""
+    test_id = uuid4()
+
+    # Mock Top DataFrame with scan dates
+    mock_top_df = pd.DataFrame({
+        "VARIETY": ["Test Variety", "Test Variety", "Test Variety"],
+        "SCAN DATE": ["2024-02-10", "2024-02-20", "2024-03-05"],
+    })
+
+    def read_excel_side_effect(file_path, sheet_name, **kwargs):
+        if sheet_name == "PPEW":
+            raise ValueError("PPEW worksheet not found")
+        elif sheet_name == "Top":
+            return mock_top_df
+        raise ValueError(f"Worksheet {sheet_name} not found")
+
+    mock_read_excel.side_effect = read_excel_side_effect
+
+    # Call the function
+    start_date, end_date = extract_dates_from_indoor_spreadsheet(
+        "/fake/path/test.xlsx", test_id
+    )
+
+    # Should use min scan_date as start_date
+    assert start_date is not None
+    assert end_date is not None
+    assert start_date.day == 10
+    assert end_date.day == 5
+
+
+@patch("app.tasks.upload_tasks.pd.read_excel")
+def test_extract_dates_missing_top_only_returns_planting_date(mock_read_excel):
+    """Test that only start_date is returned when Top worksheet is missing."""
+    test_id = uuid4()
+
+    # Mock PPEW DataFrame with planting date
+    mock_ppew_df = pd.DataFrame({
+        "VARIETY": ["Test Variety"],
+        "PI": ["12345"],
[…]

[…]
