### 2026-01-08 9f9d567
finish adding teams and project members to indoor projects, fix permission level for get requests
