### 2026-01-08 8cda50f
automatically refresh project detail form values on successful upload
