### 2026-01-12 db50b00
fix issue with labels not displaying while upload column expanded
