### 2026-01-13 303f8f8
use correct permission check for indoor project static files
