### 2026-01-15 gdslab/data-to-science#299 @morrowcj
@minyj283, Thanks for the thorough response. The addition of Author Contributions looks great!
