### 2026-01-29 #9497 @jGaboardi
@twinklejoshi – During the [pre-review stage](https://github.com/openjournals/joss-reviews/issues/8595#issuecomment-3568254192) you indicated that you would be willing to review this submission. Is there any chance I could still take you up on that offer?
