### 2026-02-06 #9497 @jGaboardi
@jbousquin – We have lost a reviewer here. Considering your potential willingness[^1] during the pre-review, would you be willing to give a review now? [One review](https://github.com/openjournals/joss-reviews/issues/9497#issuecomment-3567066653) is already complete.

[^1]: https://github.com/openjournals/joss-reviews/issues/8595#issuecomment-3553085976
