### 2026-02-08 66af1c6
document team and project member role permissions
