+++ docs/ROLE_PERMISSIONS.md
@@ -1,0 +1,352 @@
+# Role-Based Access Control (RBAC) Documentation
+
+This document summarizes what each role can do across the Data to Science platform.
+
+## Role Hierarchy
+
+The platform uses a three-tier role hierarchy for both **Team Members** and **Project Members**:
+
+| Role | Level | Description |
+|------|-------|-------------|
+| **Owner** | 3 (highest) | Full control including delete operations |
+| **Manager** | 2 | Read and write access, but cannot delete |
+| **Viewer** | 1 (lowest) | Read-only access |
+
+---
+
+## Team Permissions
+
+Teams are used to organize users and can be associated with projects.
+
+### Team Operations
+
+| Action | Owner | Manager | Viewer |
+|--------|:-----:|:-------:|:------:|
+| Create team | Any approved user | Any approved user | Any approved user |
+| View team details | Yes | Yes | Yes |
+| List teams user belongs to | Yes | Yes | Yes |
+| Update team | Yes | No | No |
+| Delete team | Yes | No | No |
+
+### Team Member Management
+
+| Action | Owner | Manager | Viewer |
+|--------|:-----:|:-------:|:------:|
+| List team members | Yes | Yes | Yes |
+| Add team member | Yes | No | No |
+| Add multiple team members | Yes | No | No |
+| Update team member role | Yes | Yes* | No |
+| Remove team member | Yes | No | No |
+
+*\*Managers can update roles but **cannot** modify owners or promote members to owner.*
+
+**Special Restrictions:**
+- The team creator (original owner) cannot be modified or removed
+- Demo users cannot perform write/delete operations on teams
+
+---
+
+## Project Permissions
+
+Projects are the primary organizational unit containing flights, data products, and other resources.
+
+### Project Operations
+
+| Action | Owner | Manager | Viewer |
+|--------|:-----:|:-------:|:------:|
+| Create project | Any approved user | Any approved user | Any approved user |
+| View project | Yes | Yes | Yes |
+| List user's projects | Yes | Yes | Yes |
+| Update project | Yes | Yes | No |
+| Deactivate (delete) project | Yes | No | No |
+| Bookmark/unbookmark project | Yes | Yes | Yes |
+
+### Project Member Management
+
+| Action | Owner | Manager | Viewer |
+|--------|:-----:|:-------:|:------:|
+| List project members | Yes | Yes | Yes |
+| View project member | Yes | Yes | Yes |
+| Add project member | Yes | No | No |
+| Add multiple project members | Yes | No | No |
+| Update project member role | Yes | No | No |
+| Remove project member | Yes | No | No |
+
+**Special Restrictions:**
+- The project creator (original owner) cannot be removed
+- Demo users cannot perform write/delete operations on projects
+
+---
+
+## Flight Permissions
+
+Flights belong to projects and contain raw data and data products.
+
+### Flight Operations
+
+| Action | Owner | Manager | Viewer |
+|--------|:-----:|:-------:|:------:|
+| Create flight | Yes | Yes | No |
+| View flight | Yes | Yes | Yes |
+| List flights in project | Yes | Yes | Yes |
+| Update flight | Yes | Yes | No |
+| Move flight to another project | Yes* | No | No |
+| Deactivate (delete) flight | Yes | Yes** | No |
+| Check flight processing progress | Yes | Yes | No |
+
+*\*User must be Owner of both source and destination projects.*
+
+*\*\*Flight deletion requires `can_read_write_flight` AND `can_read_write_delete_project` - effectively requiring Manager+ on flight and Owner on project.*
+
+**Additional Restriction:**
[…]
