### 2026-02-13 #9497 @jGaboardi
👋 @genematx, @wybert, @goergen95 - Would one or more of you be willing to review this submission for JOSS? We've already got [one complete review](https://github.com/openjournals/joss-reviews/issues/9497#issuecomment-3567066653) and need 1 more! We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
