### 2026-02-20 78061f8
update redis image and harden container configuration
