### 2026-02-20 4590330
add explicit build configs and pin image versions across compose files
