+++ docker-compose.prod.example.yml
@@ -128,6 +128,12 @@ services:
       - no-new-privileges=true
 
   celery_worker:
+    build:
+      context: ./backend
+      dockerfile: backend.dockerfile
+      args:
+        NUM_OF_WORKERS: 1
+        LIMIT_MAX_REQUESTS: 10000
     image: d2s-api:latest
     platform: linux/x86_64
     command: /app/celery/worker-start.sh
@@ -163,6 +169,12 @@ services:
       - db.env
 
   celery_beat:
+    build:
+      context: ./backend
+      dockerfile: backend.dockerfile
+      args:
+        NUM_OF_WORKERS: 1
+        LIMIT_MAX_REQUESTS: 10000
     image: d2s-api:latest
     platform: linux/x86_64
     command: /app/celery/beat-start.sh
@@ -250,7 +262,7 @@ services:
       - VARNISH_SIZE=2G
 
   pg_tileserv:
-    image: pramsey/pg_tileserv:latest
+    image: pramsey/pg_tileserv:20250131-amd64
     platform: linux/x86_64
     volumes:
       - ./pg_tileserv.toml:/config/pg_tileserv.toml:ro
@@ -271,7 +283,7 @@ services:
       - no-new-privileges=true
 
   titiler:
-    image: ghcr.io/developmentseed/titiler:0.23.1
+    image: ghcr.io/developmentseed/titiler:1.2.0
     volumes:
       - user-data:/static
     environment:
@@ -332,7 +344,7 @@ services:
       - no-new-privileges=true
 
   tusd:
-    image: tusproject/tusd:latest
+    image: tusproject/tusd:sha-0e52ad6
     command: -behind-proxy -hooks-http http://proxy/api/v1/tusd -hooks-http-forward-headers Cookie X-Project-ID X-Flight-ID X-Data-Type
     volumes:
       - tusd-data:/srv/tusd-data/data
