+++ backend/app/middleware.py
@@ -1,5 +1,7 @@
+import asyncio
 import logging
 import time
+from datetime import datetime, timezone
 from typing import Awaitable, Callable, Optional
 from uuid import UUID
 
@@ -14,6 +16,14 @@ from app.core.security import decode_token
 
 logger = logging.getLogger(__name__)
 
+# In-memory cache of user_id -> last time a DB activity update was scheduled.
+# Prevents a background task (and DB connection) from being queued on every
+# single request when the throttle window hasn't elapsed yet.
+_activity_cache: dict[UUID, datetime] = {}
+
+# How long (seconds) to wait for the sync DB work before giving up.
+_ACTIVITY_TRACKING_TIMEOUT = 5.0
+
 
 def write_access_log(request: Request, response: Response, process_time: float) -> None:
     """Write HTTP access log entry."""
@@ -25,8 +35,8 @@ def write_access_log(request: Request, response: Response, process_time: float) 
     )
 
 
-def track_activity_background(user_id: UUID) -> None:
-    """Background task to update user's last activity timestamp."""
+def _sync_track_activity(user_id: UUID) -> None:
+    """Synchronous DB work for updating user activity. Runs in a threadpool."""
     db_session = None
     try:
         db_session = next(get_db())
@@ -38,11 +48,44 @@ def track_activity_background(user_id: UUID) -> None:
                 throttle_minutes=settings.ACTIVITY_TRACKING_THROTTLE_MINUTES,
             )
     except Exception as e:
-        # Silently log errors during activity tracking
         logger.debug(f"Error tracking user activity: {e}")
     finally:
         if db_session:
             db_session.close()
+
+
+async def track_activity_background(user_id: UUID) -> None:
+    """
+    Async background task that runs the sync DB activity update in a threadpool
+    with a hard timeout so it can never block a Uvicorn worker from restarting.
+    """
+    try:
+        await asyncio.wait_for(
+            asyncio.to_thread(_sync_track_activity, user_id),
+            timeout=_ACTIVITY_TRACKING_TIMEOUT,
+        )
+    except asyncio.TimeoutError:
+        logger.debug(
+            f"Activity tracking timed out after {_ACTIVITY_TRACKING_TIMEOUT}s "
+            f"for user {user_id}"
+        )
+    except Exception as e:
+        logger.debug(f"Error tracking user activity: {e}")
+
+
+def _should_track_activity(user_id: UUID) -> bool:
+    """
+    Return True only if enough time has passed since we last scheduled a DB
+    update for this user. Uses an in-memory cache so we don't queue a task
+    (and open a DB connection) on every single request.
+    """
+    now = datetime.now(timezone.utc)
+    last = _activity_cache.get(user_id)
+    throttle_seconds = settings.ACTIVITY_TRACKING_THROTTLE_MINUTES * 60
+    if last is None or (now - last).total_seconds() > throttle_seconds:
+        _activity_cache[user_id] = now
+        return True
+    return False
 
 
 async def log_and_track_middleware(
@@ -52,17 +95,18 @@ async def log_and_track_middleware(
     Combined middleware to log HTTP requests and track user activity.
 
     Performs two functions:
-    1. Logs HTTP requests with timing information
-    2. Tracks user activity for authenticated requests (with throttling)
+    1. Logs HTTP requests with timing information.
+    2. Tracks user activity for authenticated requests (with throttling).
 
     Both tasks run as background tasks after the response is sent to avoid
-    interfering with request handling.
+    interfering with request handling. Activity tracking is gated by an
+    in-memory cache so a DB connection is only opened once per throttle window
+    per user, regardless of request volume.
     """
     # Extract user info before processing request
-    user_id_to_track = None
+    user_id_to_track: Optional[UUID] = None
 
     try:
-        # Check for JWT token in cookies
[…]
