### 2026-03-05 66dee20
switch raw data progress tracking from RPC polling to push-based updates
