### 2026-03-05 1d8dc1d
patch JobManager db session in tests to use test fixture session
