### 2026-03-06 d432619
Add d2s_admin CLI for quickstart deployment and superuser creation
