### 2026-03-07 934f535
suppress warnings, more feedback during superuser creation
