### 2026-03-07 47d974c
Add email validation, clean CLI output, and default HTTP_COOKIE_SECURE to 0
