### 2026-03-08 286a636
Use --wait flag on quickstart up to ensure all services are healthy before returning
