### 2026-03-09 cbd2766
reorganize readme and split into multiple docs
