+++ docs/README.md
@@ -4,4 +4,6 @@ Welcome! Here's what's available:
 
 - [System Architecture](architecture_summary.md)
 - [Backend Folder Structure](backend_structure.md)
+- [Configuration Options](options.md)
+- [Local Development Guide](local_development.md)
 - [JOSS Paper](paper.md)
