+++ docs/options.md
@@ -1,0 +1,88 @@
+# Configuration Options
+
+D2S uses environment files to configure its services. This document describes all available environment variables organized by file.
+
+## `.env`
+
+| Variable | Description |
+|---|---|
+| `EXTERNAL_STORAGE` | Location where raw image zips and metadata will be sent for image processing jobs. It could be a mapped network drive or any other directory on the host machine. This **should be left empty** unless you have set up an image processing backend that works with the D2S image processing Celery task. |
+| `TUSD_STORAGE` | Location of Docker managed volume or mapped host directory that stores user uploaded datasets. |
+| `TILE_SIGNING_SECRET` | Secret key used for creating a signed URL that the client can use to access raster tiles and MVT tiles. |
+
+## `frontend.env`
+
+| Variable | Description |
+|---|---|
+| `VITE_MAPBOX_ACCESS_TOKEN` | Mapbox access token for satellite imagery (optional). |
+| `VITE_MAPTILER_API_KEY` | Maptiler API key for OSM labels (optional). |
+| `VITE_TURNSTILE_SITE_KEY` | Cloudflare Turnstile site key for bot protection on registration (optional). Leave empty to disable. Must be paired with `TURNSTILE_SECRET_KEY` in `backend.env`. |
+
+## `backend.env`
+
+You must provide a value for `SECRET_KEY` in your `backend.env` file. Use a cryptographically secure random string of at least 32 characters.
+
+| Variable | Description |
+|---|---|
+| `API_PROJECT_NAME` | Name that will appear in the FastAPI docs. |
+| `API_DOMAIN` | Domain used for accessing the application (e.g., `http://localhost` or `https://customdomain`). |
+| `CELERY_BROKER_URL` | Address for local redis service. |
+| `CELERY_RESULT_BACKEND` | Address for local redis service. |
+| `ENABLE_BREEDBASE` | Enable/disable Breedbase connection endpoints (`true`/`false`). |
+| `ENABLE_CAMPAIGNS` | Enable/disable campaign management endpoints (`true`/`false`). |
+| `ENABLE_IFORESTER` | Enable/disable iForester integration endpoints (`true`/`false`). |
+| `ENABLE_STAC` | Enable/disable STAC (SpatioTemporal Asset Catalog) endpoints (`true`/`false`). |
+| `ENABLE_OPENTELEMETRY` | Enable/disable OpenTelemetry. Must also uncomment the `otel-collector` container and toggle the `backend` and `titiler` OpenTelemetry related environment settings in the docker compose config. Disabled by default. |
+| `EXTENSIONS` | Can be used to enable extensions. Should be left blank typically. |
+| `EXTERNAL_STORAGE` | Internal mount point for external storage. Should be blank unless you have a binding mount for external storage. |
+| `EXTERNAL_VIEWER_URL` | Web application for displaying published STAC Items (optional). |
+| `MAIL_ENABLED` | Enable SMTP email by changing value from `0` to `1`. |
+| `MAIL_SERVER` | SMTP server address. |
+| `MAIL_USERNAME` | Username for SMTP server. |
+| `MAIL_PASSWORD` | Password for SMTP server. |
+| `MAIL_FROM` | Sender email address. |
+| `MAIL_FROM_NAME` | Name of sender. |
+| `MAIL_ADMINS` | List of emails that should receive admin mail separated by commas. |
+| `MAIL_PORT` | SMTP server port. |
+| `MAPBOX_ACCESS_TOKEN` | Mapbox access token for satellite imagery (optional). |
[…]
