+++ README.md
@@ -9,6 +9,17 @@
   <img src="docs/assets/d2s-screenshot4.png" width="45%" />
 </p>
 
+## Table of Contents
+
+- [What is D2S?](#what-is-d2s)
+- [What Makes D2S Unique?](#what-makes-d2s-unique)
+- [What Can You Do with D2S?](#what-can-you-do-with-d2s)
+- [Deploying D2S](#-deploying-d2s)
+  - [Quick Start](#-quick-start)
+  - [Local Development](#-local-development)
+- [Example Deployment](#example-deployment)
+- [Documentation](#-documentation)
+
 ## What is D2S?
 
 The Data to Science (D2S) platform is an innovative, open-source initiative designed to facilitate data sharing and collaboration among researchers worldwide. Developed by Jinha Jung, an associate professor of civil engineering at Purdue University, and his team, the platform primarily focuses on housing data from unmanned aerial vehicles (UAVs) used in agricultural and forestry research.
@@ -55,16 +66,23 @@ Additional documentation and tutorials are available for users who want to explo
 - **User manual:** https://docs.gdsl.org/data-to-science-user-manual
 - **Website and tutorial videos:** https://d2s.org
 
-## ⚡ Quick Start
+## 🚀 Deploying D2S
 
-### 📋 Prerequisites
+D2S offers two paths for deployment:
+
+- **Quick Start** — Uses prebuilt images and the `d2s_admin` CLI to get a running instance with minimal setup. Best for trying out the platform.
+- **Local Development** — Builds from source with full environment configuration. Best for contributors and custom deployments.
+
+### ⚡ Quick Start
+
+#### 📋 Prerequisites
 
 - [Python 3](https://www.python.org/) (standard library only — no additional packages required)
 - [Docker Engine](https://docs.docker.com/engine/install/) and [Docker Compose](https://docs.docker.com/compose/install/)
 
 If you can successfully run `python3 --version`, `docker --version`, and `docker compose --version` from a terminal then you are ready to proceed.
 
-### Initialize and start
+#### Initialize and start
 
 From the root directory of the repository, run the following commands to set up and start the quickstart deployment:
 
@@ -75,7 +93,7 @@ python3 -m d2s_admin quickstart up
 
 The `init` command copies the example environment files and creates the `tusd-data/` upload directory. The `up` command starts all service containers in the background.
 
-### 👤 Create a superuser
+#### 👤 Create a superuser
 
 Once the containers are running, create an admin account:
 
@@ -89,7 +107,7 @@ You will be prompted for an email, first name, last name, and password. You can 
 python3 -m d2s_admin createsuperuser --email admin@example.com --first-name Admin --last-name User --password yourpassword
 ```
 
-### ⏹️ Stop the containers
+#### ⏹️ Stop the containers
 
 ```
 python3 -m d2s_admin quickstart down
@@ -121,7 +139,7 @@ If you prefer to set up manually instead of using `d2s_admin`:
 
 </details>
 
-### 🌍 Accessing the web application
+#### 🌍 Accessing the web application
 
 The Data To Science web application can be accessed from `http://localhost:8000`. It may take up to a minute for the backend to finish initializing. If you are running D2S on a virtual machine or remote server and accessing it via a LAN IP over HTTP, update `HTTP_COOKIE_SECURE` in your `backend.env` file to:
 
@@ -131,208 +149,9 @@ HTTP_COOKIE_SECURE=0
 
 This allows cookies to work correctly in non-localhost HTTP environments.
 
-## ⚙️ Getting started with local code
+### 🛠️ Local Development
 
-### 📋 Prerequisites
-
[…]

[…]
