### 2026-03-09 998c136
Add Companion Tools section to README
