### 2026-03-12 01a9bca
Restructure docs with MkDocs Material and GitHub Pages deployment
