+++ docs/tutorials/index.md
@@ -1,0 +1,24 @@
+# Tutorials
+
+Tutorials are **learning-oriented** lessons that guide you through using D2S. For detailed, step-by-step instructions on working with the platform, see the resources below.
+
+## User Manual
+
+The [D2S User Manual](https://docs.gdsl.org/data-to-science-user-manual/) provides comprehensive documentation for platform usage, including creating projects, uploading data, managing teams, and visualizing data products.
+
+!!! note "Under review"
+    The user manual is currently undergoing a review and update to reflect the latest features available in D2S. Some sections may not yet cover recent additions.
+
+## Training Workshop Videos
+
+The [D2S website](https://d2s.org/) hosts training workshop videos with detailed instructions for common workflows. These videos walk through real-world use cases in agricultural and forestry research.
+
+To request access to additional training materials, contact [Data to Science Support](mailto:support@d2s.org).
+
+## Getting Started with a Local Instance
+
+If you want to deploy and run D2S locally, the [Getting Started](quickstart.md) tutorial below walks you through setting up a local instance using Docker.
+
+| Tutorial                         | What you'll learn                        |
+| -------------------------------- | ---------------------------------------- |
+| [Getting Started](quickstart.md) | Set up a local D2S instance using Docker |
