+++ docs/reference/roles-permissions.md
@@ -1,8 +1,8 @@
-# Role-Based Access Control (RBAC) Documentation
+# Roles and Permissions
 
-This document summarizes what each role can do across the Data to Science platform.
+This document summarizes what each role can do across the D2S platform. For the design rationale behind this system, see [Authentication and Authorization](../explanation/authentication.md).
 
-## Role Hierarchy
+## Role hierarchy
 
 The platform uses a three-tier role hierarchy for both **Team Members** and **Project Members**:
 
@@ -14,11 +14,11 @@ The platform uses a three-tier role hierarchy for both **Team Members** and **Pr
 
 ---
 
-## Team Permissions
+## Team permissions
 
 Teams are used to organize users and can be associated with projects.
 
-### Team Operations
+### Team operations
 
 | Action | Owner | Manager | Viewer |
 |--------|:-----:|:-------:|:------:|
@@ -28,7 +28,7 @@ Teams are used to organize users and can be associated with projects.
 | Update team | Yes | No | No |
 | Delete team | Yes | No | No |
 
-### Team Member Management
+### Team member management
 
 | Action | Owner | Manager | Viewer |
 |--------|:-----:|:-------:|:------:|
@@ -40,17 +40,18 @@ Teams are used to organize users and can be associated with projects.
 
 *\*Managers can update roles but **cannot** modify owners or promote members to owner.*
 
-**Special Restrictions:**
+**Special restrictions:**
+
 - The team creator (original owner) cannot be modified or removed
 - Demo users cannot perform write/delete operations on teams
 
 ---
 
-## Project Permissions
+## Project permissions
 
 Projects are the primary organizational unit containing flights, data products, and other resources.
 
-### Project Operations
+### Project operations
 
 | Action | Owner | Manager | Viewer |
 |--------|:-----:|:-------:|:------:|
@@ -61,7 +62,7 @@ Projects are the primary organizational unit containing flights, data products, 
 | Deactivate (delete) project | Yes | No | No |
 | Bookmark/unbookmark project | Yes | Yes | Yes |
 
-### Project Member Management
+### Project member management
 
 | Action | Owner | Manager | Viewer |
 |--------|:-----:|:-------:|:------:|
@@ -72,17 +73,18 @@ Projects are the primary organizational unit containing flights, data products, 
 | Update project member role | Yes | No | No |
 | Remove project member | Yes | No | No |
 
-**Special Restrictions:**
+**Special restrictions:**
+
 - The project creator (original owner) cannot be removed
 - Demo users cannot perform write/delete operations on projects
 
 ---
 
-## Flight Permissions
+## Flight permissions
 
 Flights belong to projects and contain raw data and data products.
 
-### Flight Operations
+### Flight operations
 
 | Action | Owner | Manager | Viewer |
 |--------|:-----:|:-------:|:------:|
@@ -96,18 +98,19 @@ Flights belong to projects and contain raw data and data products.
 
 *\*User must be Owner of both source and destination projects.*
 
-*\*\*Flight deletion requires `can_read_write_flight` AND `can_read_write_delete_project` - effectively requiring Manager+ on flight and Owner on project.*
+*\*\*Flight deletion requires `can_read_write_flight` AND `can_read_write_delete_project` — effectively requiring Manager+ on flight and Owner on project.*
 
-**Additional Restriction:**
+**Additional restriction:**
+
 - Flights cannot be deactivated when the project is published to a STAC catalog
 
 ---
 
-## Data Product Permissions
+## Data product permissions
 
[…]

[…]
