### 2026-03-12 a5132ab
Add platform compatibility notes, STAC catalog links, and fix doc references
