+++ docs/README.md
@@ -1,23 +1,0 @@
-# Project Documentation
-
-This documentation is organized using the [Diátaxis](https://diataxis.fr/) framework into four sections:
-
-- **[Tutorials](tutorials/index.md)** — Learning-oriented guides for getting started
-- **[How-To Guides](how-to/index.md)** — Task-oriented recipes for specific goals
-- **[Reference](reference/index.md)** — Technical specifications (API, configuration, data models)
-- **[Explanation](explanation/index.md)** — Design rationale and architecture discussion
-
-## Building the docs locally
-
-```bash
-pip install mkdocs-material
-mkdocs serve
-```
-
-Then open [http://localhost:8000](http://localhost:8000) in your browser.
-
-## Other resources
-
-- [JOSS Paper](paper.md)
-- [User Manual](https://docs.gdsl.org/data-to-science-user-manual/)
-- [D2S Website](https://d2s.org/)
