+++ docs/tutorials/quickstart.md
@@ -17,6 +17,9 @@ docker --version
 docker compose --version
 ```
 
+!!! note "Platform support"
+    D2S containers target the `linux/x86_64` architecture. If you are running Docker Desktop on Apple Silicon (ARM), emulation is handled automatically, but you may notice slower build times and startup.
+
 ## Step 1: Initialize the environment
 
 From the root directory of the repository, run:
