### 2026-03-17 #9497 @jGaboardi
@minyj283 – please review the *Additional Author Tasks After Review is Complete* [above](https://github.com/openjournals/joss-reviews/issues/9497#issuecomment-4076638911) for those 5 bullet points.

Once you have completed those bullets, ping me here and I will start on the *Editor Tasks Prior to Acceptance*.
