### 2026-03-18 #9497 @minyj283
Hi, @jGaboardi : We have double-checked authors and affiliations, and **v1.1.0** is the version with the latest changes from the reviewers. The DOI number for the Zenodo archive is **10.5281/zenodo.19073184**, ([LINK](https://doi.org/10.5281/zenodo.19073184)) of which the author list and the license are identical to this manuscript.
