### 2026-03-18 #9497 @jGaboardi
@minyj283 – Please check the title cases for all references and have them match what's listed at the official DOI.

For example for `(Breunig et al., 2020)` the title casing should be *Geospatial Data Management Research: Progress and Future Directions* not *Geospatial data management research: Progress and future directions*.
