### 2026-03-18 7b49292
Update titles in bibliography entries for consistency
