+++ docs/paper.bib
@@ -1,5 +1,5 @@
 @article{breunig2020geospatial,
-  title={Geospatial Data Management Research: Progress and Future Directions},
+  title={{Geospatial Data Management Research: Progress and Future Directions}},
   author={Breunig, Martin and Bradley, Patrick Erik and Jahn, Markus and Kuper, Paul and Mazroob, Nima and R{\"o}sch, Norbert and Al-Doori, Mulhim and Stefanakis, Emmanuel and Jadidi, Mojgan},
   journal={ISPRS International Journal of Geo-Information},
   volume={9},
@@ -11,7 +11,7 @@
 }
 
 @inproceedings{mohd2018remote,
-  title={Remote sensing UAV/drones and its applications for urban areas: a review},
+  title={{Remote sensing UAV/drones and its applications for urban areas: a review}},
   author={Mohd Noor, Norzailawati and Abdullah, Alias and Hashim, Mazlan},
   booktitle={IOP conference series: Earth and environmental science},
   volume={169},
@@ -22,7 +22,7 @@
 }
 
 @article{duarte2022recent,
-  title={Recent Advances in Forest Insect Pests and Diseases Monitoring Using UAV-Based Data: A Systematic Review},
+  title={{Recent Advances in Forest Insect Pests and Diseases Monitoring Using UAV-Based Data: A Systematic Review}},
   author={Duarte, Andr{\'e} and Borralho, Nuno and Cabral, Pedro and Caetano, M{\'a}rio},
   journal={Forests},
   volume={13},
@@ -34,7 +34,7 @@
 }
 
 @article{ecke2022uav,
-  title={UAV-Based Forest Health Monitoring: A Systematic Review},
+  title={{UAV-Based Forest Health Monitoring: A Systematic Review}},
   author={Ecke, Simon and Dempewolf, Jan and Frey, Julian and Schwaller, Andreas and Endres, Ewald and Klemmt, Hans-Joachim and Tiede, Dirk and Seifert, Thomas},
   journal={Remote Sensing},
   volume={14},
@@ -46,7 +46,7 @@
 }
 
 @article{molina2023review,
-  title={A Review of Unmanned Aerial Vehicle Applications in Construction Management: 2016--2021},
+  title={{A Review of Unmanned Aerial Vehicle Applications in Construction Management: 2016--2021}},
   author={Molina, Andres Acero and Huang, Yilei and Jiang, Yuhan},
   journal={Standards},
   volume={3},
@@ -58,7 +58,7 @@
 }
 
 @article{wu2020geemap,
-  title={geemap: A Python package for interactive mapping with Google Earth Engine},
+  title={{geemap: A Python package for interactive mapping with Google Earth Engine}},
   author={Wu, Qiusheng},
   journal={Journal of Open Source Software},
   volume={5},
@@ -69,7 +69,7 @@
 }
 
 @article{wu2021leafmap,
-  title={Leafmap: A Python package for interactive mapping and geospatial analysis with minimal coding in a Jupyter environment},
+  title={{Leafmap: A Python package for interactive mapping and geospatial analysis with minimal coding in a Jupyter environment}},
   author={Wu, Qiusheng},
   journal={Journal of Open Source Software},
   volume={6},
