### 2026-03-18 #9497 @minyj283
I have added an extra set of brackets. It should work properly now.
