### 2026-03-19 #9497 @kthyng
@minyj283 Paper suggestions:

- **Page 1 – Summary**
  Missing noun after “within the D2S.”
  - Current: `...archived data within the D2S and other external geospatial datasets...`
  - Suggested: `...archived data within the D2S ecosystem and other external geospatial datasets...`

- **Page 2 – Features table**
  Compound modifier missing hyphen.
  - Current: `User configurable settings for the ODM pipeline`
  - Suggested: `User-configurable settings for the ODM pipeline`

- **Page 3 – Author Contributions**
  Capitalization mid-sentence.
  - Current: `...funding acquisition, J.J.; All authors have read...`
  - Suggested: `...funding acquisition, J.J.; all authors have read...`
